@@ -14,19 +14,20 @@
 
 #include "rmw/allocators.h"
 #include "rmw/error_handling.h"
-#include "rmw/rmw.h"
 #include "rmw/impl/cpp/macros.hpp"
+#include "rmw/rmw.h"
 
 #include "rmw_gurumdds_cpp/dds_include.hpp"
 #include "rmw_gurumdds_cpp/identifier.hpp"
-#include "rmw_gurumdds_cpp/rmw_wait.hpp"
+#include "rmw_gurumdds_cpp/event_info_common.hpp"
+#include "rmw_gurumdds_cpp/wait.hpp"
 
 extern "C"
 {
 rmw_wait_set_t *
 rmw_create_wait_set(rmw_context_t * context, size_t max_conditions)
 {
-  RCUTILS_CHECK_ARGUMENT_FOR_NULL(context, NULL);
+  RCUTILS_CHECK_ARGUMENT_FOR_NULL(context, nullptr);
   RMW_CHECK_FOR_NULL_WITH_MSG(
     context->impl,
     "expected initialized context",
@@ -37,10 +38,10 @@ rmw_create_wait_set(rmw_context_t * context, size_t max_conditions)
     RMW_GURUMDDS_ID,
     return nullptr);
 
-  (void)max_conditions;
+  RCUTILS_UNUSED(max_conditions);
   rmw_wait_set_t * wait_set = rmw_wait_set_allocate();
 
-  GurumddsWaitSetInfo * wait_set_info = nullptr;
+  rmw_gurumdds_cpp::WaitSetInfo * wait_set_info = nullptr;
 
   if (!wait_set) {
     RMW_SET_ERROR_MSG("failed to allocate wait set");
@@ -48,8 +49,8 @@ rmw_create_wait_set(rmw_context_t * context, size_t max_conditions)
   }
 
   wait_set->implementation_identifier = RMW_GURUMDDS_ID;
-  wait_set->data = rmw_allocate(sizeof(GurumddsWaitSetInfo));
-  wait_set_info = static_cast<GurumddsWaitSetInfo *>(wait_set->data);
+  wait_set->data = rmw_allocate(sizeof(rmw_gurumdds_cpp::WaitSetInfo));
+  wait_set_info = static_cast<rmw_gurumdds_cpp::WaitSetInfo *>(wait_set->data);
 
   if (!wait_set_info) {
     RMW_SET_ERROR_MSG("failed to allocate wait set");
@@ -114,7 +115,7 @@ rmw_destroy_wait_set(rmw_wait_set_t * wait_set)
     RMW_GURUMDDS_ID,
     return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
 
-  GurumddsWaitSetInfo * wait_set_info = static_cast<GurumddsWaitSetInfo *>(wait_set->data);
+  rmw_gurumdds_cpp::WaitSetInfo * wait_set_info = static_cast<rmw_gurumdds_cpp::WaitSetInfo *>(wait_set->data);
 
   if (wait_set_info->active_conditions != nullptr) {
     dds_ConditionSeq_delete(wait_set_info->active_conditions);
@@ -151,7 +152,7 @@ rmw_wait(
   rmw_wait_set_t * wait_set,
   const rmw_time_t * wait_timeout)
 {
-  return __rmw_wait<GurumddsSubscriberInfo, GurumddsServiceInfo, GurumddsClientInfo>(
+  return rmw_gurumdds_cpp::wait(
     RMW_GURUMDDS_ID, subscriptions, guard_conditions,
     services, clients, events, wait_set, wait_timeout);
 }
