@@ -12,38 +12,40 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef RMW_GURUMDDS_CPP__DEMANGLE_HPP_
-#define RMW_GURUMDDS_CPP__DEMANGLE_HPP_
+#ifndef RMW_GURUMDDS__DEMANGLE_HPP_
+#define RMW_GURUMDDS__DEMANGLE_HPP_
 
 #include <string>
 
+namespace rmw_gurumdds_cpp
+{
 std::string
-_demangle_if_ros_topic(const std::string & topic_name);
+demangle_if_ros_topic(const std::string & topic_name);
 
 std::string
-_demangle_if_ros_type(const std::string & dds_type_string);
+demangle_if_ros_type(const std::string & dds_type_string);
 
 std::string
-_demangle_ros_topic_from_topic(const std::string & topic_name);
+demangle_ros_topic_from_topic(const std::string & topic_name);
 
 std::string
-_demangle_service_from_topic(const std::string & topic_name);
+demangle_service_from_topic(const std::string & topic_name);
 
 std::string
-_demangle_service_request_from_topic(const std::string & topic_name);
+demangle_service_request_from_topic(const std::string & topic_name);
 
 std::string
-_demangle_service_reply_from_topic(const std::string & topic_name);
+demangle_service_reply_from_topic(const std::string & topic_name);
 
 std::string
-_demangle_service_type_only(const std::string & dds_type_name);
+demangle_service_type_only(const std::string & dds_type_name);
 
 // Used when ros names are not mangled.
 std::string
-_identity_demangle(const std::string & name);
-
+identity_demangle(const std::string & name);
+} // namespace rmw_gurumdds_cpp
 
 using DemangleFunction = std::string (*)(const std::string &);
 using MangleFunction = DemangleFunction;
 
-#endif  // RMW_GURUMDDS_CPP__DEMANGLE_HPP_
+#endif // RMW_GURUMDDS__DEMANGLE_HPP_
