@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef RMW_GURUMDDS_CPP__GET_ENTITIES_HPP_
-#define RMW_GURUMDDS_CPP__GET_ENTITIES_HPP_
+#ifndef RMW_GURUMDDS__GET_ENTITIES_HPP_
+#define RMW_GURUMDDS__GET_ENTITIES_HPP_
 
 #include "rmw/rmw.h"
 
@@ -22,7 +22,6 @@
 
 namespace rmw_gurumdds_cpp
 {
-
 RMW_GURUMDDS_CPP_PUBLIC
 dds_DomainParticipant *
 get_participant(rmw_node_t * node);
@@ -58,7 +57,6 @@ get_request_data_reader(rmw_service_t * service);
 RMW_GURUMDDS_CPP_PUBLIC
 dds_DataWriter *
 get_response_data_writer(rmw_service_t * service);
+} // namespace rmw_gurumdds_cpp
 
-}  // namespace rmw_gurumdds_cpp
-
-#endif  // RMW_GURUMDDS_CPP__GET_ENTITIES_HPP_
+#endif // RMW_GURUMDDS__GET_ENTITIES_HPP_
