@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef RMW_GURUMDDS_CPP__IDENTIFIER_HPP_
-#define RMW_GURUMDDS_CPP__IDENTIFIER_HPP_
+#ifndef RMW_GURUMDDS__IDENTIFIER_HPP_
+#define RMW_GURUMDDS__IDENTIFIER_HPP_
 
 extern const char * const RMW_GURUMDDS_ID;
 
-#endif  // RMW_GURUMDDS_CPP__IDENTIFIER_HPP_
+#endif // RMW_GURUMDDS__IDENTIFIER_HPP_
