@@ -19,7 +19,6 @@
 
 #include "rcutils/macros.h"
 
-#include "rmw/allocators.h"
 #include "rmw/error_handling.h"
 #include "rmw/impl/cpp/macros.hpp"
 #include "rmw/init.h"
@@ -28,13 +27,12 @@
 #include "rmw/types.h"
 
 #include "rmw_dds_common/context.hpp"
-#include "rmw_dds_common/gid_utils.hpp"
-#include "rmw_dds_common/msg/participant_entities_info.hpp"
 
 #include "rmw_gurumdds_cpp/context_listener_thread.hpp"
 #include "rmw_gurumdds_cpp/graph_cache.hpp"
 #include "rmw_gurumdds_cpp/identifier.hpp"
 #include "rmw_gurumdds_cpp/rmw_context_impl.hpp"
+#include "rmw_gurumdds_cpp/wait.hpp"
 
 static
 dds_Condition *
@@ -60,8 +58,8 @@ void rmw_gurumdds_listener_thread(rmw_context_impl_t * ctx)
 {
   RCUTILS_LOG_DEBUG_NAMED(RMW_GURUMDDS_ID, "[listener thread] starting up...");
 
-  GurumddsSubscriberInfo * sub_partinfo =
-    reinterpret_cast<GurumddsSubscriberInfo *>(ctx->common_ctx.sub->data);
+  auto sub_partinfo
+    = reinterpret_cast<rmw_gurumdds_cpp::SubscriberInfo *>(ctx->common_ctx.sub->data);
   dds_ReturnCode_t ret = dds_RETCODE_ERROR;
 
   uint32_t active_len = 0;
@@ -78,7 +76,7 @@ void rmw_gurumdds_listener_thread(rmw_context_impl_t * ctx)
   dds_GuardCondition * gcond_exit =
     reinterpret_cast<dds_GuardCondition *>(ctx->common_ctx.listener_thread_gc->data);
 
-  GurumddsWaitSetInfo * waitset_info = new(std::nothrow) GurumddsWaitSetInfo();
+  auto waitset_info = new(std::nothrow) rmw_gurumdds_cpp::WaitSetInfo();
   if (waitset_info == nullptr) {
     RMW_SET_ERROR_MSG("failed to allocate WaitSetInfo");
     goto cleanup;
@@ -143,7 +141,7 @@ void rmw_gurumdds_listener_thread(rmw_context_impl_t * ctx)
     for (uint32_t i = 0; i < active_len && active; i++) {
       cond_active = dds_ConditionSeq_get(waitset_info->active_conditions, i);
       if (nullptr != cond_part_info && cond_part_info == cond_active) {
-        graph_on_participant_info(ctx);
+        rmw_gurumdds_cpp::graph_cache::on_participant_info(ctx);
       } else {
         RMW_SET_ERROR_MSG("unexpected active condition");
         goto cleanup;
@@ -194,6 +192,8 @@ cleanup:
   RCUTILS_LOG_DEBUG_NAMED(RMW_GURUMDDS_ID, "[listener thread] done");
 }
 
+namespace rmw_gurumdds_cpp
+{
 rmw_ret_t
 run_listener_thread(rmw_context_t * ctx)
 {
@@ -226,7 +226,6 @@ run_listener_thread(rmw_context_t * ctx)
   return RMW_RET_ERROR;
 }
 
-
 rmw_ret_t
 stop_listener_thread(rmw_context_t * ctx)
 {
@@ -255,3 +254,4 @@ stop_listener_thread(rmw_context_t * ctx)
   RCUTILS_LOG_DEBUG_NAMED(RMW_GURUMDDS_ID, "[listener thread] stopped");
   return RMW_RET_OK;
 }
+} //namespace rmw_gurumdds_cpp
