@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <utility>
-
 #include "rmw_gurumdds_cpp/event_converter.hpp"
 
 /// mapping of RMW_EVENT to the corresponding dds_StatusKind.
@@ -22,11 +20,17 @@ static const dds_StatusKind g_mask_map[] {
   dds_REQUESTED_DEADLINE_MISSED_STATUS,  // RMW_EVENT_REQUESTED_DEADLINE_MISSED
   dds_REQUESTED_INCOMPATIBLE_QOS_STATUS,  // RMW_EVENT_REQUESTED_QOS_INCOMPATIBLE
   dds_SAMPLE_LOST_STATUS,  // RMW_EVENT_MESSAGE_LOST
+  dds_INCONSISTENT_TOPIC_STATUS, // RMW_EVENT_SUBSCRIPTION_INCOMPATIBLE_TYPE
+  dds_SUBSCRIPTION_MATCHED_STATUS, // RMW_EVENT_SUBSCRIPTION_MATCHED
   dds_LIVELINESS_LOST_STATUS,  // RMW_EVENT_LIVELINESS_LOST
   dds_OFFERED_DEADLINE_MISSED_STATUS,  // RMW_EVENT_OFFERED_DEADLINE_MISSED
-  dds_OFFERED_INCOMPATIBLE_QOS_STATUS  // RMW_EVENT_OFFERED_QOS_INCOMPATIBLE
+  dds_OFFERED_INCOMPATIBLE_QOS_STATUS,  // RMW_EVENT_OFFERED_QOS_INCOMPATIBLE
+  dds_INCONSISTENT_TOPIC_STATUS, // RMW_EVENT_PUBLISHER_INCOMPATIBLE_TYPE
+  dds_PUBLICATION_MATCHED_STATUS // RMW_EVENT_PUBLICATION_MATCHED
 };
 
+namespace rmw_gurumdds_cpp
+{
 dds_StatusKind get_status_kind_from_rmw(const rmw_event_type_t event_t)
 {
   if (!is_event_supported(event_t)) {
@@ -38,18 +42,20 @@ dds_StatusKind get_status_kind_from_rmw(const rmw_event_type_t event_t)
 
 bool is_event_supported(const rmw_event_type_t event_t)
 {
-  return 0 <= event_t && event_t < RMW_EVENT_INVALID;
+  return event_t < RMW_EVENT_INVALID;
 }
 
 rmw_ret_t check_dds_ret_code(const dds_ReturnCode_t dds_return_code)
 {
-  if (dds_return_code == dds_RETCODE_OK) {
-    return RMW_RET_OK;
-  } else if (dds_return_code == dds_RETCODE_ERROR) {
-    return RMW_RET_ERROR;
-  } else if (dds_return_code == dds_RETCODE_TIMEOUT) {
-    return RMW_RET_TIMEOUT;
-  } else {
-    return RMW_RET_ERROR;
+  switch(dds_return_code) {
+    case dds_RETCODE_OK:
+      return RMW_RET_OK;
+    case dds_RETCODE_ERROR:
+      return RMW_RET_ERROR;
+    case dds_RETCODE_TIMEOUT:
+      return RMW_RET_TIMEOUT;
+    default:
+      return RMW_RET_ERROR;
   }
 }
+} // namespace rmw_gurumdds_cpp
