@@ -13,32 +13,31 @@
 // limitations under the License.
 
 #include <chrono>
-#include <limits>
-#include <random>
 #include <string>
-#include <thread>
 #include <utility>
 
-#include "rcutils/logging_macros.h"
 #include "rcutils/error_handling.h"
+#include "rcutils/logging_macros.h"
 
 #include "rmw/allocators.h"
-#include "rmw/rmw.h"
 #include "rmw/error_handling.h"
 #include "rmw/impl/cpp/macros.hpp"
+#include "rmw/rmw.h"
 #include "rmw/types.h"
 #include "rmw/validate_full_topic_name.h"
 
-#include "rmw_gurumdds_cpp/gid.hpp"
+#include "rmw_dds_common/qos.hpp"
+
+#include "rmw_gurumdds_cpp/event_converter.hpp"
 #include "rmw_gurumdds_cpp/graph_cache.hpp"
 #include "rmw_gurumdds_cpp/identifier.hpp"
 #include "rmw_gurumdds_cpp/names_and_types_helpers.hpp"
 #include "rmw_gurumdds_cpp/namespace_prefix.hpp"
 #include "rmw_gurumdds_cpp/qos.hpp"
 #include "rmw_gurumdds_cpp/rmw_context_impl.hpp"
-#include "rmw_gurumdds_cpp/types.hpp"
-
-#include "type_support_service.hpp"
+#include "rmw_gurumdds_cpp/gid.hpp"
+#include "rmw_gurumdds_cpp/type_support_service.hpp"
+#include "rmw_gurumdds_cpp/event_info_common.hpp"
 
 extern "C"
 {
@@ -54,7 +53,7 @@ rmw_create_client(
     node,
     node->implementation_identifier,
     RMW_GURUMDDS_ID,
-    return nullptr)
+    return nullptr);
   RMW_CHECK_ARGUMENT_FOR_NULL(type_supports, nullptr);
   RMW_CHECK_ARGUMENT_FOR_NULL(service_name, nullptr);
   if (strlen(service_name) == 0) {
@@ -63,7 +62,11 @@ rmw_create_client(
   }
   RMW_CHECK_ARGUMENT_FOR_NULL(qos_policies, nullptr);
 
-  if (!qos_policies->avoid_ros_namespace_conventions) {
+   // Adapt any 'best available' QoS options
+  rmw_qos_profile_t adapted_qos_policies =
+    rmw_dds_common::qos_profile_update_best_available_for_services(*qos_policies);
+
+  if (!adapted_qos_policies.avoid_ros_namespace_conventions) {
     int validation_result = RMW_TOPIC_VALID;
     rmw_ret_t ret = rmw_validate_full_topic_name(service_name, &validation_result, nullptr);
     if (ret != RMW_RET_OK) {
@@ -92,7 +95,7 @@ rmw_create_client(
   rmw_context_impl_t * ctx = node->context->impl;
   std::lock_guard<std::mutex> guard(ctx->endpoint_mutex);
 
-  GurumddsClientInfo * client_info = nullptr;
+  rmw_gurumdds_cpp::ClientInfo * client_info = nullptr;
   rmw_client_t * rmw_client = nullptr;
 
   dds_DomainParticipant * participant = ctx->participant;
@@ -104,6 +107,10 @@ rmw_create_client(
 
   dds_DataWriter * request_writer = nullptr;
   dds_DataReader * response_reader = nullptr;
+  dds_DataReaderListener response_listener;
+  dds_DataSeq* data_seq = nullptr;
+  dds_SampleInfoSeq* info_seq = nullptr;
+  dds_UnsignedLongSeq* raw_data_sizes = nullptr;
   dds_ReadCondition * read_condition = nullptr;
   dds_TypeSupport * request_typesupport = nullptr;
   dds_TypeSupport * response_typesupport = nullptr;
@@ -123,10 +130,11 @@ rmw_create_client(
   std::string response_type_name;
   std::string request_metastring;
   std::string response_metastring;
+  const rosidl_type_hash_t* type_hash;
 
   // Create topic and type name strings
   service_type_name =
-    create_service_type_name(type_support->data, type_support->typesupport_identifier);
+    rmw_gurumdds_cpp::create_service_type_name(type_support->data, type_support->typesupport_identifier);
   request_type_name = service_type_name.first;
   response_type_name = service_type_name.second;
   if (request_type_name.empty() || response_type_name.empty()) {
@@ -136,30 +144,19 @@ rmw_create_client(
 
   request_topic_name.reserve(256);
   response_topic_name.reserve(256);
-  request_topic_name = create_topic_name(
-    ros_service_requester_prefix, service_name, "Request", qos_policies);
-  response_topic_name = create_topic_name(
-    ros_service_response_prefix, service_name, "Reply", qos_policies);
+  request_topic_name = rmw_gurumdds_cpp::create_topic_name(
+    rmw_gurumdds_cpp::ros_service_requester_prefix, service_name, "Request", &adapted_qos_policies);
+  response_topic_name = rmw_gurumdds_cpp::create_topic_name(
+    rmw_gurumdds_cpp::ros_service_response_prefix, service_name, "Reply", &adapted_qos_policies);
 
   service_metastring =
-    create_service_metastring(type_support->data, type_support->typesupport_identifier);
+    rmw_gurumdds_cpp::create_service_metastring(type_support->data, type_support->typesupport_identifier);
   request_metastring = service_metastring.first;
   response_metastring = service_metastring.second;
   if (request_metastring.empty() || response_metastring.empty()) {
     RMW_SET_ERROR_MSG("failed to create metastring");
     return nullptr;
   }
-
-  client_info = new(std::nothrow) GurumddsClientInfo();
-  if (client_info == nullptr) {
-    RMW_SET_ERROR_MSG("failed to allocate GurumddsClientInfo");
-    goto fail;
-  }
-
-  client_info->implementation_identifier = RMW_GURUMDDS_ID;
-  client_info->service_typesupport = type_support;
-  client_info->sequence_number = 0;
-  client_info->ctx = ctx;
 
   request_typesupport = dds_TypeSupport_create(request_metastring.c_str());
   if (request_typesupport == nullptr) {
@@ -262,7 +259,8 @@ rmw_create_client(
   }
 
   // Create datawriter for request
-  if (!get_datawriter_qos(publisher, qos_policies, &datawriter_qos)) {
+  type_hash = type_support->request_typesupport->get_type_hash_func(type_support->request_typesupport);
+  if (!rmw_gurumdds_cpp::get_datawriter_qos(publisher, &adapted_qos_policies, *type_hash, &datawriter_qos)) {
     // Error message already set
     goto fail;
   }
@@ -274,7 +272,6 @@ rmw_create_client(
     dds_DataWriterQos_finalize(&datawriter_qos);
     goto fail;
   }
-  client_info->request_writer = request_writer;
 
   ret = dds_DataWriterQos_finalize(&datawriter_qos);
   if (ret != dds_RETCODE_OK) {
@@ -282,7 +279,8 @@ rmw_create_client(
     goto fail;
   }
 
-  if (!get_datareader_qos(subscriber, qos_policies, &datareader_qos)) {
+  type_hash = type_support->response_typesupport->get_type_hash_func(type_support->response_typesupport);
+  if (!rmw_gurumdds_cpp::get_datareader_qos(subscriber, &adapted_qos_policies, *type_hash, &datareader_qos)) {
     // error message already set
     goto fail;
   }
@@ -294,7 +292,6 @@ rmw_create_client(
     dds_DataReaderQos_finalize(&datareader_qos);
     goto fail;
   }
-  client_info->response_reader = response_reader;
 
   ret = dds_DataReaderQos_finalize(&datareader_qos);
   if (ret != dds_RETCODE_OK) {
@@ -308,16 +305,59 @@ rmw_create_client(
     RMW_SET_ERROR_MSG("failed to create read condition");
     goto fail;
   }
+
+  client_info = new(std::nothrow) rmw_gurumdds_cpp::ClientInfo();
+  if (client_info == nullptr) {
+    RMW_SET_ERROR_MSG("failed to allocate ClientInfo");
+    goto fail;
+  }
+
+  data_seq = dds_DataSeq_create(1);
+  if (nullptr == data_seq) {
+    RMW_SET_ERROR_MSG("failed to allocate data_seq");
+    return nullptr;
+  }
+  info_seq = dds_SampleInfoSeq_create(1);
+  if (nullptr == info_seq) {
+    RMW_SET_ERROR_MSG("failed to allocate info_seq");
+    return nullptr;
+  }
+  raw_data_sizes = dds_UnsignedLongSeq_create(1);
+  if (nullptr == raw_data_sizes) {
+    RMW_SET_ERROR_MSG("failed to allocate raw_data_sizes");
+    return nullptr;
+  }
+
+  dds_DataReader_set_listener_context(response_reader, client_info);
+  response_listener.on_data_available = [](const dds_DataReader * response_reader){
+    dds_DataReader* reader = const_cast<dds_DataReader*>(response_reader);
+    auto* info = static_cast<rmw_gurumdds_cpp::ClientInfo*>(dds_DataReader_get_listener_context(reader));
+    std::lock_guard<std::mutex> guard(info->event_callback_data.mutex);
+    if(info->event_callback_data.callback) {
+      info->event_callback_data.callback(info->event_callback_data.user_data, info->count_unread());
+    }
+  };
+
+  client_info->request_writer = request_writer;
+  client_info->response_reader = response_reader;
   client_info->read_condition = read_condition;
+  client_info->response_listener = response_listener;
+  client_info->data_seq = data_seq;
+  client_info->info_seq = info_seq;
+  client_info->raw_data_sizes = raw_data_sizes;
+  client_info->implementation_identifier = RMW_GURUMDDS_ID;
+  client_info->service_typesupport = type_support;
+  client_info->sequence_number = 0;
+  client_info->ctx = ctx;
 
   // Set GUID
   dds_DataWriter_get_guid(request_writer, client_guid);
-  memcpy(client_info->writer_guid, client_guid, sizeof(client_guid));
+  std::memcpy(client_info->writer_guid, client_guid, sizeof(client_guid));
 
-  entity_get_gid(
+  rmw_gurumdds_cpp::entity_get_gid(
     reinterpret_cast<dds_Entity *>(client_info->request_writer),
     client_info->publisher_gid);
-  entity_get_gid(
+  rmw_gurumdds_cpp::entity_get_gid(
     reinterpret_cast<dds_Entity *>(client_info->response_reader),
     client_info->subscriber_gid);
 
@@ -326,7 +366,7 @@ rmw_create_client(
     RMW_SET_ERROR_MSG("failed to allocate memory for client");
     goto fail;
   }
-  memset(rmw_client, 0, sizeof(rmw_client_t));
+  std::memset(rmw_client, 0, sizeof(rmw_client_t));
 
   rmw_client->implementation_identifier = RMW_GURUMDDS_ID;
   rmw_client->data = client_info;
@@ -335,9 +375,9 @@ rmw_create_client(
     RMW_SET_ERROR_MSG("failed to allocate memory for client name");
     goto fail;
   }
-  memcpy(const_cast<char *>(rmw_client->service_name), service_name, strlen(service_name) + 1);
+  std::memcpy(const_cast<char *>(rmw_client->service_name), service_name, strlen(service_name) + 1);
 
-  if (graph_on_client_created(ctx, node, client_info) != RMW_RET_OK) {
+  if (rmw_gurumdds_cpp::graph_cache::on_client_created(ctx, node, client_info) != RMW_RET_OK) {
     RCUTILS_LOG_ERROR_NAMED(RMW_GURUMDDS_ID, "failed to update graph for client creation");
     goto fail;
   }
@@ -416,7 +456,7 @@ rmw_destroy_client(rmw_node_t * node, rmw_client_t * client)
   rmw_context_impl_t * ctx = node->context->impl;
   std::lock_guard<std::mutex> guard(ctx->endpoint_mutex);
 
-  GurumddsClientInfo * client_info = static_cast<GurumddsClientInfo *>(client->data);
+  auto client_info = static_cast<rmw_gurumdds_cpp::ClientInfo *>(client->data);
 
   if (client_info != nullptr) {
     if (client_info->request_writer != nullptr) {
@@ -426,6 +466,10 @@ rmw_destroy_client(rmw_node_t * node, rmw_client_t * client)
         return RMW_RET_ERROR;
       }
     }
+
+    dds_DataSeq_delete(client_info->data_seq);
+    dds_SampleInfoSeq_delete(client_info->info_seq);
+    dds_UnsignedLongSeq_delete(client_info->raw_data_sizes);
 
     if (client_info->response_reader != nullptr) {
       if (client_info->read_condition != nullptr) {
@@ -443,7 +487,7 @@ rmw_destroy_client(rmw_node_t * node, rmw_client_t * client)
       }
     }
 
-    if (graph_on_client_deleted(ctx, node, client_info) != RMW_RET_OK) {
+    if (rmw_gurumdds_cpp::graph_cache::on_client_deleted(ctx, node, client_info) != RMW_RET_OK) {
       RCUTILS_LOG_ERROR_NAMED(RMW_GURUMDDS_ID, "failed to update graph for client deletion");
       return RMW_RET_ERROR;
     }
@@ -485,7 +529,7 @@ rmw_service_server_is_available(
     return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
   RMW_CHECK_ARGUMENT_FOR_NULL(is_available, RMW_RET_ERROR);
 
-  GurumddsClientInfo * client_info = static_cast<GurumddsClientInfo *>(client->data);
+  auto client_info = static_cast<rmw_gurumdds_cpp::ClientInfo *>(client->data);
   if (client_info == nullptr) {
     RMW_SET_ERROR_MSG("client info handle is null");
     return RMW_RET_ERROR;
@@ -553,6 +597,29 @@ rmw_service_server_is_available(
 }
 
 rmw_ret_t
+rmw_get_gid_for_client(const rmw_client_t * client, rmw_gid_t * gid)
+{
+  RMW_CHECK_ARGUMENT_FOR_NULL(client, RMW_RET_INVALID_ARGUMENT);
+  RMW_CHECK_ARGUMENT_FOR_NULL(gid, RMW_RET_INVALID_ARGUMENT);
+  RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
+    client,
+    client->implementation_identifier,
+    RMW_GURUMDDS_ID,
+    return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
+
+  auto client_info = static_cast<rmw_gurumdds_cpp::ClientInfo *>(client->data);
+  if (client_info == nullptr) {
+    RMW_SET_ERROR_MSG("client info is null");
+    return RMW_RET_ERROR;
+  }
+
+  *gid = client_info->publisher_gid;
+  gid->implementation_identifier = RMW_GURUMDDS_ID;
+
+  return RMW_RET_OK;
+}
+
+rmw_ret_t
 rmw_client_request_publisher_get_actual_qos(
   const rmw_client_t * client,
   rmw_qos_profile_t * qos)
@@ -565,7 +632,7 @@ rmw_client_request_publisher_get_actual_qos(
     return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
   RMW_CHECK_ARGUMENT_FOR_NULL(qos, RMW_RET_INVALID_ARGUMENT);
 
-  GurumddsClientInfo * client_info = static_cast<GurumddsClientInfo *>(client->data);
+  auto client_info = static_cast<rmw_gurumdds_cpp::ClientInfo *>(client->data);
   if (client_info == nullptr) {
     RMW_SET_ERROR_MSG("client info is null");
     return RMW_RET_ERROR;
@@ -584,13 +651,13 @@ rmw_client_request_publisher_get_actual_qos(
     return RMW_RET_ERROR;
   }
 
-  qos->reliability = convert_reliability(&dds_qos.reliability);
-  qos->durability = convert_durability(&dds_qos.durability);
-  qos->deadline = convert_deadline(&dds_qos.deadline);
-  qos->lifespan = convert_lifespan(&dds_qos.lifespan);
-  qos->liveliness = convert_liveliness(&dds_qos.liveliness);
-  qos->liveliness_lease_duration = convert_liveliness_lease_duration(&dds_qos.liveliness);
-  qos->history = convert_history(&dds_qos.history);
+  qos->reliability = rmw_gurumdds_cpp::convert_reliability(&dds_qos.reliability);
+  qos->durability = rmw_gurumdds_cpp::convert_durability(&dds_qos.durability);
+  qos->deadline = rmw_gurumdds_cpp::convert_deadline(&dds_qos.deadline);
+  qos->lifespan = rmw_gurumdds_cpp::convert_lifespan(&dds_qos.lifespan);
+  qos->liveliness = rmw_gurumdds_cpp::convert_liveliness(&dds_qos.liveliness);
+  qos->liveliness_lease_duration = rmw_gurumdds_cpp::convert_liveliness_lease_duration(&dds_qos.liveliness);
+  qos->history = rmw_gurumdds_cpp::convert_history(&dds_qos.history);
   qos->depth = static_cast<size_t>(dds_qos.history.depth);
 
   ret = dds_DataWriterQos_finalize(&dds_qos);
@@ -615,7 +682,7 @@ rmw_client_response_subscription_get_actual_qos(
     return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
   RMW_CHECK_ARGUMENT_FOR_NULL(qos, RMW_RET_INVALID_ARGUMENT);
 
-  GurumddsClientInfo * client_info = static_cast<GurumddsClientInfo *>(client->data);
+  auto client_info = static_cast<rmw_gurumdds_cpp::ClientInfo *>(client->data);
   if (client_info == nullptr) {
     RMW_SET_ERROR_MSG("client info is null");
     return RMW_RET_ERROR;
@@ -634,12 +701,12 @@ rmw_client_response_subscription_get_actual_qos(
     return RMW_RET_ERROR;
   }
 
-  qos->reliability = convert_reliability(&dds_qos.reliability);
-  qos->durability = convert_durability(&dds_qos.durability);
-  qos->deadline = convert_deadline(&dds_qos.deadline);
-  qos->liveliness = convert_liveliness(&dds_qos.liveliness);
-  qos->liveliness_lease_duration = convert_liveliness_lease_duration(&dds_qos.liveliness);
-  qos->history = convert_history(&dds_qos.history);
+  qos->reliability = rmw_gurumdds_cpp::convert_reliability(&dds_qos.reliability);
+  qos->durability = rmw_gurumdds_cpp::convert_durability(&dds_qos.durability);
+  qos->deadline = rmw_gurumdds_cpp::convert_deadline(&dds_qos.deadline);
+  qos->liveliness = rmw_gurumdds_cpp::convert_liveliness(&dds_qos.liveliness);
+  qos->liveliness_lease_duration = rmw_gurumdds_cpp::convert_liveliness_lease_duration(&dds_qos.liveliness);
+  qos->history = rmw_gurumdds_cpp::convert_history(&dds_qos.history);
   qos->depth = static_cast<size_t>(dds_qos.history.depth);
 
   ret = dds_DataReaderQos_finalize(&dds_qos);
@@ -666,7 +733,7 @@ rmw_send_request(
   RMW_CHECK_ARGUMENT_FOR_NULL(ros_request, RMW_RET_INVALID_ARGUMENT);
   RMW_CHECK_ARGUMENT_FOR_NULL(sequence_id, RMW_RET_INVALID_ARGUMENT);
 
-  GurumddsClientInfo * client_info = static_cast<GurumddsClientInfo *>(client->data);
+  auto client_info = static_cast<rmw_gurumdds_cpp::ClientInfo *>(client->data);
   if (client_info == nullptr) {
     RMW_SET_ERROR_MSG("client info handle is null");
     return RMW_RET_ERROR;
@@ -687,7 +754,7 @@ rmw_send_request(
   size_t size = 0;
 
   if (client_info->ctx->service_mapping_basic) {
-    void * dds_request = allocate_request_basic(
+    void * dds_request = rmw_gurumdds_cpp::allocate_request_basic(
       type_support->data,
       type_support->typesupport_identifier,
       ros_request,
@@ -698,7 +765,7 @@ rmw_send_request(
       return RMW_RET_ERROR;
     }
 
-    bool res = serialize_request_basic(
+    bool res = rmw_gurumdds_cpp::serialize_request_basic(
       type_support->data,
       type_support->typesupport_identifier,
       ros_request,
@@ -721,7 +788,7 @@ rmw_send_request(
     }
     free(dds_request);
   } else {
-    void * dds_request = allocate_request_enhanced(
+    void * dds_request = rmw_gurumdds_cpp::allocate_request_enhanced(
       type_support->data,
       type_support->typesupport_identifier,
       ros_request,
@@ -732,7 +799,7 @@ rmw_send_request(
       return RMW_RET_ERROR;
     }
 
-    bool res = serialize_request_enhanced(
+    bool res = rmw_gurumdds_cpp::serialize_request_enhanced(
       type_support->data,
       type_support->typesupport_identifier,
       ros_request,
@@ -747,11 +814,11 @@ rmw_send_request(
     }
 
     dds_SampleInfoEx sampleinfo_ex;
-    memset(&sampleinfo_ex, 0, sizeof(dds_SampleInfoEx));
-    ros_sn_to_dds_sn(++client_info->sequence_number, &sampleinfo_ex.seq);
-    ros_guid_to_dds_guid(
-      reinterpret_cast<int8_t *>(client_info->writer_guid),
-      reinterpret_cast<int8_t *>(&sampleinfo_ex.src_guid));
+    std::memset(&sampleinfo_ex, 0, sizeof(dds_SampleInfoEx));
+    rmw_gurumdds_cpp::ros_sn_to_dds_sn(++client_info->sequence_number, &sampleinfo_ex.seq);
+    rmw_gurumdds_cpp::ros_guid_to_dds_guid(
+      reinterpret_cast<const uint8_t *>(client_info->writer_guid),
+      reinterpret_cast<uint8_t *>(&sampleinfo_ex.src_guid));
 
     if (dds_DataWriter_raw_write_w_sampleinfoex(
         request_writer, dds_request, size, &sampleinfo_ex) != dds_RETCODE_OK)
@@ -786,7 +853,7 @@ rmw_take_response(
 
   *taken = false;
 
-  GurumddsClientInfo * client_info = static_cast<GurumddsClientInfo *>(client->data);
+  auto client_info = static_cast<rmw_gurumdds_cpp::ClientInfo *>(client->data);
   if (client_info == nullptr) {
     RMW_SET_ERROR_MSG("client info handle is null");
     return RMW_RET_ERROR;
@@ -864,7 +931,8 @@ rmw_take_response(
         int32_t sn_high = 0;
         uint32_t sn_low = 0;
         int8_t client_guid[16] = {0};
-        bool res = deserialize_response_basic(
+        dds_SampleInfoEx * sampleinfo_ex = reinterpret_cast<dds_SampleInfoEx *>(sample_info);
+        bool res = rmw_gurumdds_cpp::deserialize_response_basic(
           type_support->data,
           type_support->typesupport_identifier,
           ros_response,
@@ -884,14 +952,15 @@ rmw_take_response(
           return RMW_RET_ERROR;
         }
 
-        if (memcmp(client_info->writer_guid, client_guid, 16) == 0) {
+        if (std::memcmp(client_info->writer_guid, client_guid, RMW_GID_STORAGE_SIZE) == 0) {
           request_header->source_timestamp =
             sample_info->source_timestamp.sec * static_cast<int64_t>(1000000000) +
             sample_info->source_timestamp.nanosec;
-          // TODO(clemjh): SampleInfo doesn't contain received_timestamp
-          request_header->received_timestamp = 0;
+          request_header->received_timestamp =
+            sampleinfo_ex->reception_timestamp.sec * static_cast<int64_t>(1000000000) +
+            sampleinfo_ex->reception_timestamp.nanosec;
           request_header->request_id.sequence_number = ((int64_t)sn_high) << 32 | sn_low;
-          memcpy(request_header->request_id.writer_guid, client_guid, 16);
+          std::memcpy(request_header->request_id.writer_guid, client_guid, RMW_GID_STORAGE_SIZE);
 
           *taken = true;
         }
@@ -938,10 +1007,10 @@ rmw_take_response(
         int64_t sequence_number = 0;
         int8_t client_guid[16] = {0};
         dds_SampleInfoEx * sampleinfo_ex = reinterpret_cast<dds_SampleInfoEx *>(sample_info);
-        dds_guid_to_ros_guid(reinterpret_cast<int8_t *>(&sampleinfo_ex->src_guid), client_guid);
-        dds_sn_to_ros_sn(sampleinfo_ex->seq, &sequence_number);
+        rmw_gurumdds_cpp::dds_guid_to_ros_guid(reinterpret_cast<int8_t *>(&sampleinfo_ex->src_guid), client_guid);
+        rmw_gurumdds_cpp::dds_sn_to_ros_sn(sampleinfo_ex->seq, &sequence_number);
 
-        bool res = deserialize_response_enhanced(
+        bool res = rmw_gurumdds_cpp::deserialize_response_enhanced(
           type_support->data,
           type_support->typesupport_identifier,
           ros_response,
@@ -958,14 +1027,15 @@ rmw_take_response(
           return RMW_RET_ERROR;
         }
 
-        if (memcmp(client_info->writer_guid, client_guid, 16) == 0) {
+        if (std::memcmp(client_info->writer_guid, client_guid, RMW_GID_STORAGE_SIZE) == 0) {
           request_header->source_timestamp =
             sample_info->source_timestamp.sec * static_cast<int64_t>(1000000000) +
             sample_info->source_timestamp.nanosec;
-          // TODO(clemjh): SampleInfo doesn't contain received_timestamp
-          request_header->received_timestamp = 0;
+          request_header->received_timestamp =
+            sampleinfo_ex->reception_timestamp.sec * static_cast<int64_t>(1000000000) +
+            sampleinfo_ex->reception_timestamp.nanosec;
           request_header->request_id.sequence_number = sequence_number;
-          memcpy(request_header->request_id.writer_guid, client_guid, 16);
+          std::memcpy(request_header->request_id.writer_guid, client_guid, RMW_GID_STORAGE_SIZE);
 
           *taken = true;
         }
@@ -990,11 +1060,39 @@ rmw_client_set_on_new_response_callback(
   rmw_event_callback_t callback,
   const void * user_data)
 {
-  (void)rmw_client;
-  (void)callback;
-  (void)user_data;
+  RMW_CHECK_ARGUMENT_FOR_NULL(rmw_client, RMW_RET_INVALID_ARGUMENT);
+  RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
+    rmw_client,
+    rmw_client->implementation_identifier,
+    RMW_GURUMDDS_ID,
+    return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
+  auto client_info = static_cast<rmw_gurumdds_cpp::ClientInfo *>(rmw_client->data);
+  if (client_info == nullptr) {
+    RMW_SET_ERROR_MSG("invalid client data");
+    return RMW_RET_ERROR;
+  }
 
-  RMW_SET_ERROR_MSG("rmw_client_set_on_new_request_callback not implemented");
-  return RMW_RET_UNSUPPORTED;
+  std::lock_guard<std::mutex> guard(client_info->event_callback_data.mutex);
+  dds_StatusMask mask = dds_DataReader_get_status_changes(client_info->response_reader);
+  dds_ReturnCode_t dds_rc = dds_RETCODE_ERROR;
+
+  if (callback) {
+    size_t unread_count = client_info->count_unread();
+    if (0 < unread_count) {
+      callback(user_data, unread_count);
+    }
+
+    client_info->event_callback_data.callback = callback;
+    client_info->event_callback_data.user_data = user_data;
+    mask |= dds_DATA_AVAILABLE_STATUS;
+    dds_rc = dds_DataReader_set_listener(client_info->response_reader, &client_info->response_listener, mask);
+  } else {
+    client_info->event_callback_data.callback = nullptr;
+    client_info->event_callback_data.user_data = nullptr;
+    mask &= ~dds_DATA_AVAILABLE_STATUS;
+    dds_rc = dds_DataReader_set_listener(client_info->response_reader, &client_info->response_listener, mask);
+  }
+
+  return rmw_gurumdds_cpp::check_dds_ret_code(dds_rc);
 }
 }  // extern "C"
