@@ -14,9 +14,6 @@
 
 #include <map>
 #include <string>
-#include <vector>
-
-#include "rcutils/logging_macros.h"
 
 #include "rmw/error_handling.h"
 #include "rmw/impl/cpp/macros.hpp"
@@ -26,11 +23,10 @@
 
 #include "rmw_dds_common/context.hpp"
 
-#include "rmw_gurumdds_cpp/rmw_context_impl.hpp"
-#include "rmw_gurumdds_cpp/demangle.hpp"
 #include "rmw_gurumdds_cpp/identifier.hpp"
-#include "rmw_gurumdds_cpp/namespace_prefix.hpp"
 #include "rmw_gurumdds_cpp/names_and_types_helpers.hpp"
+#include "rmw_gurumdds_cpp/namespace_prefix.hpp"
+#include "rmw_gurumdds_cpp/rmw_context_impl.hpp"
 
 extern "C"
 {
@@ -52,16 +48,18 @@ rmw_count_publishers(
   if (RMW_RET_OK != ret) {
     return ret;
   }
+
   if (RMW_TOPIC_VALID != validation_result) {
     const char * reason = rmw_full_topic_name_validation_result_string(validation_result);
     RMW_SET_ERROR_MSG_WITH_FORMAT_STRING("topic_name argument is invalid: %s", reason);
     return RMW_RET_INVALID_ARGUMENT;
   }
+
   RMW_CHECK_ARGUMENT_FOR_NULL(count, RMW_RET_INVALID_ARGUMENT);
 
   auto common_ctx = &node->context->impl->common_ctx;
-  const std::string mangled_topic_name = create_topic_name(
-    ros_topic_prefix, topic_name, "", false);
+  const std::string mangled_topic_name = rmw_gurumdds_cpp::create_topic_name(
+    rmw_gurumdds_cpp::ros_topic_prefix, topic_name, "", false);
 
   return common_ctx->graph_cache.get_writer_count(mangled_topic_name, count);
 }
@@ -84,17 +82,85 @@ rmw_count_subscribers(
   if (RMW_RET_OK != ret) {
     return ret;
   }
+
   if (RMW_TOPIC_VALID != validation_result) {
     const char * reason = rmw_full_topic_name_validation_result_string(validation_result);
     RMW_SET_ERROR_MSG_WITH_FORMAT_STRING("topic_name argument is invalid: %s", reason);
     return RMW_RET_INVALID_ARGUMENT;
   }
+
   RMW_CHECK_ARGUMENT_FOR_NULL(count, RMW_RET_INVALID_ARGUMENT);
 
   auto common_ctx = &node->context->impl->common_ctx;
-  const std::string mangled_topic_name = create_topic_name(
-    ros_topic_prefix, topic_name, "", false);
+  const std::string mangled_topic_name = rmw_gurumdds_cpp::create_topic_name(
+    rmw_gurumdds_cpp::ros_topic_prefix, topic_name, "", false);
 
   return common_ctx->graph_cache.get_reader_count(mangled_topic_name, count);
+}
+
+rmw_ret_t
+rmw_count_clients(
+  const rmw_node_t * node,
+  const char * service_name,
+  size_t * count)
+{
+  RMW_CHECK_ARGUMENT_FOR_NULL(node, RMW_RET_INVALID_ARGUMENT);
+  RMW_CHECK_ARGUMENT_FOR_NULL(service_name, RMW_RET_INVALID_ARGUMENT);
+  RMW_CHECK_ARGUMENT_FOR_NULL(count, RMW_RET_INVALID_ARGUMENT);
+  RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
+    node,
+    node->implementation_identifier,
+    RMW_GURUMDDS_ID,
+    return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
+  int validation_result = RMW_TOPIC_VALID;
+  rmw_ret_t ret = rmw_validate_full_topic_name(service_name, &validation_result, nullptr);
+  if (RMW_RET_OK != ret) {
+    return ret;
+  }
+
+  if (RMW_TOPIC_VALID != validation_result) {
+    const char * reason = rmw_full_topic_name_validation_result_string(validation_result);
+    RMW_SET_ERROR_MSG_WITH_FORMAT_STRING("service_name argument is invalid: %s", reason);
+    return RMW_RET_INVALID_ARGUMENT;
+  }
+
+  auto common_ctx = &node->context->impl->common_ctx;
+  const std::string mangled_service_name = rmw_gurumdds_cpp::create_topic_name(
+    rmw_gurumdds_cpp::ros_service_response_prefix, service_name, "", false);
+
+  return common_ctx->graph_cache.get_reader_count(mangled_service_name, count);
+}
+
+rmw_ret_t
+rmw_count_services(
+  const rmw_node_t * node,
+  const char * service_name,
+  size_t * count)
+{
+  RMW_CHECK_ARGUMENT_FOR_NULL(node, RMW_RET_INVALID_ARGUMENT);
+  RMW_CHECK_ARGUMENT_FOR_NULL(service_name, RMW_RET_INVALID_ARGUMENT);
+  RMW_CHECK_ARGUMENT_FOR_NULL(count, RMW_RET_INVALID_ARGUMENT);
+  RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
+    node,
+    node->implementation_identifier,
+    RMW_GURUMDDS_ID,
+    return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
+  int validation_result = RMW_TOPIC_VALID;
+  rmw_ret_t ret = rmw_validate_full_topic_name(service_name, &validation_result, nullptr);
+  if (RMW_RET_OK != ret) {
+    return ret;
+  }
+
+  if (RMW_TOPIC_VALID != validation_result) {
+    const char * reason = rmw_full_topic_name_validation_result_string(validation_result);
+    RMW_SET_ERROR_MSG_WITH_FORMAT_STRING("service_name argument is invalid: %s", reason);
+    return RMW_RET_INVALID_ARGUMENT;
+  }
+
+  auto common_ctx = &node->context->impl->common_ctx;
+  const std::string mangled_service_name = rmw_gurumdds_cpp::create_topic_name(
+    rmw_gurumdds_cpp::ros_service_response_prefix, service_name, "", false);
+
+  return common_ctx->graph_cache.get_writer_count(mangled_service_name, count);
 }
 }  // extern "C"
