@@ -17,5 +17,11 @@
 bool
 rmw_feature_supported(rmw_feature_t feature)
 {
-  return feature == RMW_FEATURE_MESSAGE_INFO_PUBLICATION_SEQUENCE_NUMBER ? true : false;
+  switch(feature) {
+    case RMW_FEATURE_MESSAGE_INFO_RECEPTION_SEQUENCE_NUMBER:
+    case RMW_FEATURE_MESSAGE_INFO_PUBLICATION_SEQUENCE_NUMBER:
+      return true;
+    default:
+      return false;
+  }
 }
