@@ -14,7 +14,6 @@
 
 #include "rmw/error_handling.h"
 #include "rmw/get_network_flow_endpoints.h"
-#include "rmw/rmw.h"
 #include "rmw/types.h"
 
 extern "C"
@@ -25,9 +24,9 @@ rmw_publisher_get_network_flow_endpoints(
   rcutils_allocator_t * allocator,
   rmw_network_flow_endpoint_array_t * network_flow_endpoint_array)
 {
-  (void) publisher;
-  (void) allocator;
-  (void) network_flow_endpoint_array;
+  RCUTILS_UNUSED(publisher);
+  RCUTILS_UNUSED(allocator);
+  RCUTILS_UNUSED(network_flow_endpoint_array);
   RMW_SET_ERROR_MSG("rmw_publisher_get_network_flow_endpoint not implemented");
   return RMW_RET_UNSUPPORTED;
 }
@@ -38,9 +37,9 @@ rmw_subscription_get_network_flow_endpoints(
   rcutils_allocator_t * allocator,
   rmw_network_flow_endpoint_array_t * network_flow_endpoint_array)
 {
-  (void) subscription;
-  (void) allocator;
-  (void) network_flow_endpoint_array;
+  RCUTILS_UNUSED(subscription);
+  RCUTILS_UNUSED(allocator);
+  RCUTILS_UNUSED(network_flow_endpoint_array);
   RMW_SET_ERROR_MSG("rmw_subscription_get_network_flow_endpoint not implemented");
   return RMW_RET_UNSUPPORTED;
 }
