@@ -13,13 +13,7 @@
 // limitations under the License.
 
 #include <array>
-#include <utility>
-#include <set>
-#include <string>
-#include <vector>
-#include <list>
-#include <thread>
-#include <chrono>
+#include <cstring>
 
 #include "rcutils/filesystem.h"
 #include "rcutils/logging_macros.h"
@@ -29,29 +23,55 @@
 
 #include "rmw/allocators.h"
 #include "rmw/error_handling.h"
-#include "rmw/rmw.h"
 #include "rmw/impl/cpp/macros.hpp"
-#include "rmw/impl/cpp/key_value.hpp"
+#include "rmw/rmw.h"
 #include "rmw/sanity_checks.h"
-#include "rmw/convert_rcutils_ret_to_rmw_ret.h"
 #include "rmw/validate_namespace.h"
 #include "rmw/validate_node_name.h"
 
-#include "rmw_dds_common/context.hpp"
-
-#include "rmw_gurumdds_cpp/identifier.hpp"
-#include "rmw_gurumdds_cpp/dds_include.hpp"
 #include "rmw_gurumdds_cpp/graph_cache.hpp"
+#include "rmw_gurumdds_cpp/identifier.hpp"
 #include "rmw_gurumdds_cpp/rmw_context_impl.hpp"
 
-rmw_node_t *
-__rmw_create_node(
+static inline rmw_ret_t
+get_node_names(
   const char * implementation_identifier,
+  const rmw_node_t * node,
+  rcutils_string_array_t * node_names,
+  rcutils_string_array_t * node_namespaces,
+  rcutils_string_array_t * enclaves)
+{
+  RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
+    node,
+    node->implementation_identifier, implementation_identifier,
+    return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
+  rcutils_allocator_t allocator = rcutils_get_default_allocator();
+  auto common_ctx = &node->context->impl->common_ctx;
+  return common_ctx->graph_cache.get_node_names(
+    node_names,
+    node_namespaces,
+    enclaves,
+    &allocator);
+}
+
+extern "C"
+{
+rmw_node_t *
+rmw_create_node(
   rmw_context_t * context,
   const char * name,
   const char * namespace_)
 {
-  /* Validate node's name and namespace */
+  RCUTILS_CHECK_ARGUMENT_FOR_NULL(context, nullptr);
+  RMW_CHECK_FOR_NULL_WITH_MSG(
+    context->impl,
+    "expected initialized context",
+    return nullptr);
+  RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
+    context,
+    context->implementation_identifier,
+    RMW_GURUMDDS_ID,
+    return nullptr);
   int validation_result = RMW_NODE_NAME_VALID;
   rmw_ret_t ret = rmw_validate_node_name(name, &validation_result, nullptr);
   if (RMW_RET_OK != ret) {
@@ -121,7 +141,7 @@ __rmw_create_node(
     RMW_SET_ERROR_MSG("failed to allocate memory for node name");
     return nullptr;
   }
-  memcpy(const_cast<char *>(node_handle->name), name, strlen(name) + 1);
+  std::memcpy(const_cast<char *>(node_handle->name), name, strlen(name) + 1);
 
   node_handle->namespace_ =
     static_cast<const char *>(rmw_allocate(sizeof(char) * strlen(namespace_) + 1));
@@ -129,13 +149,13 @@ __rmw_create_node(
     RMW_SET_ERROR_MSG("failed to allocate memory for node namespace");
     return nullptr;
   }
-  memcpy(const_cast<char *>(node_handle->namespace_), namespace_, strlen(namespace_) + 1);
+  std::memcpy(const_cast<char *>(node_handle->namespace_), namespace_, strlen(namespace_) + 1);
 
-  node_handle->implementation_identifier = implementation_identifier;
+  node_handle->implementation_identifier = RMW_GURUMDDS_ID;
   node_handle->data = nullptr;
   node_handle->context = context;
 
-  if (graph_on_node_created(ctx, node_handle) != RMW_RET_OK) {
+  if (rmw_gurumdds_cpp::graph_cache::on_node_created(ctx, node_handle) != RMW_RET_OK) {
     RMW_SET_ERROR_MSG("failed to create node");
     return nullptr;
   }
@@ -150,18 +170,18 @@ __rmw_create_node(
 }
 
 rmw_ret_t
-__rmw_destroy_node(const char * implementation_identifier, rmw_node_t * node)
+rmw_destroy_node(rmw_node_t * node)
 {
   RMW_CHECK_ARGUMENT_FOR_NULL(node, RMW_RET_INVALID_ARGUMENT);
   RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
     node,
-    node->implementation_identifier, implementation_identifier,
+    node->implementation_identifier, RMW_GURUMDDS_ID,
     return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
 
   rmw_context_impl_t * ctx = node->context->impl;
   std::lock_guard<std::mutex> guard(ctx->initialization_mutex);
 
-  if (graph_on_node_deleted(ctx, node) != RMW_RET_OK) {
+  if (rmw_gurumdds_cpp::graph_cache::on_node_deleted(ctx, node) != RMW_RET_OK) {
     RMW_SET_ERROR_MSG("failed to update for node delete");
     return RMW_RET_ERROR;
   }
@@ -183,105 +203,18 @@ __rmw_destroy_node(const char * implementation_identifier, rmw_node_t * node)
 }
 
 const rmw_guard_condition_t *
-__rmw_node_get_graph_guard_condition(const rmw_node_t * node)
-{
-  auto common_ctx = &node->context->impl->common_ctx;
-  if (!common_ctx) {
-    RMW_SET_ERROR_MSG("common_context is nullptr");
-    return nullptr;
-  }
-  return common_ctx->graph_guard_condition;
-}
-
-rmw_ret_t
-_get_node_names(
-  const char * implementation_identifier,
-  const rmw_node_t * node,
-  rcutils_string_array_t * node_names,
-  rcutils_string_array_t * node_namespaces,
-  rcutils_string_array_t * enclaves)
-{
-  RMW_CHECK_ARGUMENT_FOR_NULL(node, RMW_RET_INVALID_ARGUMENT);
-  RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
-    node,
-    node->implementation_identifier, implementation_identifier,
-    return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
-  if (rmw_check_zero_rmw_string_array(node_names) != RMW_RET_OK ||
-    rmw_check_zero_rmw_string_array(node_namespaces) != RMW_RET_OK)
-  {
-    return RMW_RET_INVALID_ARGUMENT;
-  }
-
-  if (enclaves != nullptr &&
-    rmw_check_zero_rmw_string_array(enclaves) != RMW_RET_OK)
-  {
-    return RMW_RET_INVALID_ARGUMENT;
-  }
-
-  auto common_ctx = &node->context->impl->common_ctx;
-  rcutils_allocator_t allocator = rcutils_get_default_allocator();
-
-  return common_ctx->graph_cache.get_node_names(
-    node_names,
-    node_namespaces,
-    enclaves,
-    &allocator);
-}
-
-rmw_ret_t
-__rmw_get_node_names(
-  const char * implementation_identifier,
-  const rmw_node_t * node,
-  rcutils_string_array_t * node_names,
-  rcutils_string_array_t * node_namespaces)
-{
-  return _get_node_names(implementation_identifier, node, node_names, node_namespaces, nullptr);
-}
-
-rmw_ret_t
-__rmw_get_node_names_with_enclaves(
-  const char * implementation_identifier,
-  const rmw_node_t * node,
-  rcutils_string_array_t * node_names,
-  rcutils_string_array_t * node_namespaces,
-  rcutils_string_array_t * enclaves)
-{
-  return _get_node_names(
-    implementation_identifier, node, node_names, node_namespaces, enclaves);
-}
-
-extern "C"
-{
-rmw_node_t *
-rmw_create_node(
-  rmw_context_t * context,
-  const char * name,
-  const char * namespace_)
-{
-  RCUTILS_CHECK_ARGUMENT_FOR_NULL(context, nullptr);
-  RMW_CHECK_FOR_NULL_WITH_MSG(
-    context->impl,
-    "expected initialized context",
-    return nullptr);
-  RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
-    context,
-    context->implementation_identifier,
-    RMW_GURUMDDS_ID,
-    return nullptr);
-  return __rmw_create_node(
-    RMW_GURUMDDS_ID, context, name, namespace_);
-}
-
-rmw_ret_t
-rmw_destroy_node(rmw_node_t * node)
-{
-  return __rmw_destroy_node(RMW_GURUMDDS_ID, node);
-}
-
-const rmw_guard_condition_t *
 rmw_node_get_graph_guard_condition(const rmw_node_t * node)
 {
-  return __rmw_node_get_graph_guard_condition(node);
+  RMW_CHECK_FOR_NULL_WITH_MSG(
+    node,
+    "expected initialized node",
+    return nullptr);
+  auto common_ctx = &node->context->impl->common_ctx;
+  RMW_CHECK_FOR_NULL_WITH_MSG(
+    common_ctx,
+    "expected initialized common_ctx",
+    return nullptr);
+  return common_ctx->graph_guard_condition;
 }
 
 rmw_ret_t
@@ -290,8 +223,7 @@ rmw_get_node_names(
   rcutils_string_array_t * node_names,
   rcutils_string_array_t * node_namespaces)
 {
-  return __rmw_get_node_names(
-    RMW_GURUMDDS_ID, node, node_names, node_namespaces);
+  return get_node_names(RMW_GURUMDDS_ID, node, node_names, node_namespaces, nullptr);
 }
 
 rmw_ret_t
@@ -301,7 +233,6 @@ rmw_get_node_names_with_enclaves(
   rcutils_string_array_t * node_namespaces,
   rcutils_string_array_t * enclaves)
 {
-  return __rmw_get_node_names_with_enclaves(
-    RMW_GURUMDDS_ID, node, node_names, node_namespaces, enclaves);
+  return get_node_names(RMW_GURUMDDS_ID, node, node_names, node_namespaces, enclaves);
 }
 }  // extern "C"
