@@ -12,38 +12,27 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <chrono>
 #include <cstring>
-#include <functional>
-#include <map>
 #include <set>
 #include <string>
-#include <thread>
-#include <vector>
 
 #include "rcutils/allocator.h"
 #include "rcutils/error_handling.h"
 #include "rcutils/logging_macros.h"
-#include "rcutils/strdup.h"
-#include "rcutils/types.h"
 
 #include "rmw/allocators.h"
-#include "rmw/convert_rcutils_ret_to_rmw_ret.h"
 #include "rmw/error_handling.h"
 #include "rmw/get_node_info_and_types.h"
-#include "rmw/get_topic_names_and_types.h"
 #include "rmw/impl/cpp/macros.hpp"
-#include "rmw/impl/cpp/key_value.hpp"
 #include "rmw/names_and_types.h"
-#include "rmw/rmw.h"
+#include "rmw/validate_namespace.h"
+#include "rmw/validate_node_name.h"
 
 #include "rmw_dds_common/context.hpp"
 
 #include "rmw_gurumdds_cpp/dds_include.hpp"
 #include "rmw_gurumdds_cpp/demangle.hpp"
-#include "rmw_gurumdds_cpp/guid.hpp"
 #include "rmw_gurumdds_cpp/identifier.hpp"
-#include "rmw_gurumdds_cpp/names_and_types_helpers.hpp"
 #include "rmw_gurumdds_cpp/rmw_context_impl.hpp"
 
 using GetNamesAndTypesByNodeFunction = rmw_ret_t (*)(
@@ -55,26 +44,8 @@ using GetNamesAndTypesByNodeFunction = rmw_ret_t (*)(
   rcutils_allocator_t *,
   rmw_names_and_types_t *);
 
-rmw_ret_t
-validate_names_and_namespace(
-  const char * node_name,
-  const char * node_namespace)
-{
-  if (node_name == nullptr) {
-    RMW_SET_ERROR_MSG("node name is null");
-    return RMW_RET_INVALID_ARGUMENT;
-  }
-
-  if (node_namespace == nullptr) {
-    RMW_SET_ERROR_MSG("node namespace is null");
-    return RMW_RET_INVALID_ARGUMENT;
-  }
-
-  return RMW_RET_OK;
-}
-
-static rmw_ret_t
-__get_topic_names_and_types_by_node(
+static inline rmw_ret_t
+get_topic_names_and_types_by_node(
   const rmw_node_t * node,
   rcutils_allocator_t * allocator,
   const char * node_name,
@@ -93,24 +64,35 @@ __get_topic_names_and_types_by_node(
     return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
   RCUTILS_CHECK_ALLOCATOR_WITH_MSG(
     allocator, "allocator argument is invalid", return RMW_RET_INVALID_ARGUMENT);
-
-  rmw_ret_t ret = rmw_names_and_types_check_zero(topic_names_and_types);
-  if (ret != RMW_RET_OK) {
+  int validation_result = RMW_NODE_NAME_VALID;
+  rmw_ret_t ret = rmw_validate_node_name(node_name, &validation_result, nullptr);
+  if (RMW_RET_OK != ret) {
     return ret;
   }
-
-  ret = validate_names_and_namespace(node_name, node_namespace);
-  if (ret != RMW_RET_OK) {
+  if (RMW_NODE_NAME_VALID != validation_result) {
+    const char * reason = rmw_node_name_validation_result_string(validation_result);
+    RMW_SET_ERROR_MSG_WITH_FORMAT_STRING("node_name argument is invalid: %s", reason);
+    return RMW_RET_INVALID_ARGUMENT;
+  }
+  validation_result = RMW_NAMESPACE_VALID;
+  ret = rmw_validate_namespace(node_namespace, &validation_result, nullptr);
+  if (RMW_RET_OK != ret) {
     return ret;
   }
-
+  if (RMW_NAMESPACE_VALID != validation_result) {
+    const char * reason = rmw_namespace_validation_result_string(validation_result);
+    RMW_SET_ERROR_MSG_WITH_FORMAT_STRING("node_namespace argument is invalid: %s", reason);
+    return RMW_RET_INVALID_ARGUMENT;
+  }
+  ret = rmw_names_and_types_check_zero(topic_names_and_types);
+  if (RMW_RET_OK != ret) {
+    return ret;
+  }
   auto common_ctx = &node->context->impl->common_ctx;
-
   if (no_demangle) {
-    demangle_topic = _identity_demangle;
-    demangle_type = _identity_demangle;
+    demangle_topic = rmw_gurumdds_cpp::identity_demangle;
+    demangle_type = rmw_gurumdds_cpp::identity_demangle;
   }
-
   return get_names_and_types_by_node(
     common_ctx,
     node_name,
@@ -121,8 +103,8 @@ __get_topic_names_and_types_by_node(
     topic_names_and_types);
 }
 
-static rmw_ret_t
-__get_reader_names_and_types_by_node(
+static inline rmw_ret_t
+get_reader_names_and_types_by_node(
   rmw_dds_common::Context * common_context,
   const std::string & node_name,
   const std::string & node_namespace,
@@ -140,8 +122,8 @@ __get_reader_names_and_types_by_node(
     topic_names_and_types);
 }
 
-static rmw_ret_t
-__get_writer_names_and_types_by_node(
+static inline rmw_ret_t
+get_writer_names_and_types_by_node(
   rmw_dds_common::Context * common_context,
   const std::string & node_name,
   const std::string & node_namespace,
@@ -175,15 +157,15 @@ rmw_get_subscriber_names_and_types_by_node(
     "rmw_get_subscriber_names_and_types_by_node: "
     "node=%s%s, demangle=%s",
     node_namespace, node_name, no_demangle == true ? "false" : "true");
-  return __get_topic_names_and_types_by_node(
+  return get_topic_names_and_types_by_node(
     node,
     allocator,
     node_name,
     node_namespace,
-    _demangle_ros_topic_from_topic,
-    _demangle_if_ros_type,
+    rmw_gurumdds_cpp::demangle_ros_topic_from_topic,
+    rmw_gurumdds_cpp::demangle_if_ros_type,
     no_demangle,
-    __get_reader_names_and_types_by_node,
+    get_reader_names_and_types_by_node,
     topic_names_and_types);
 }
 
@@ -201,15 +183,15 @@ rmw_get_publisher_names_and_types_by_node(
     "rmw_get_publisher_names_and_types_by_node: "
     "node=%s%s, demangle=%s",
     node_namespace, node_name, no_demangle == true ? "false" : "true");
-  return __get_topic_names_and_types_by_node(
+  return get_topic_names_and_types_by_node(
     node,
     allocator,
     node_name,
     node_namespace,
-    _demangle_ros_topic_from_topic,
-    _demangle_if_ros_type,
+    rmw_gurumdds_cpp::demangle_ros_topic_from_topic,
+    rmw_gurumdds_cpp::demangle_if_ros_type,
     no_demangle,
-    __get_writer_names_and_types_by_node,
+    get_writer_names_and_types_by_node,
     topic_names_and_types);
 }
 
@@ -225,15 +207,15 @@ rmw_get_service_names_and_types_by_node(
     RMW_GURUMDDS_ID,
     "rmw_get_service_names_and_types_by_node: "
     "node=%s%s", node_namespace, node_name);
-  return __get_topic_names_and_types_by_node(
+  return get_topic_names_and_types_by_node(
     node,
     allocator,
     node_name,
     node_namespace,
-    _demangle_service_request_from_topic,
-    _demangle_service_type_only,
+    rmw_gurumdds_cpp::demangle_service_request_from_topic,
+    rmw_gurumdds_cpp::demangle_service_type_only,
     false,
-    __get_reader_names_and_types_by_node,
+    get_reader_names_and_types_by_node,
     service_names_and_types);
 }
 
@@ -249,15 +231,15 @@ rmw_get_client_names_and_types_by_node(
     RMW_GURUMDDS_ID,
     "rmw_get_client_names_and_types_by_node: "
     "node=%s%s", node_namespace, node_name);
-  return __get_topic_names_and_types_by_node(
+  return get_topic_names_and_types_by_node(
     node,
     allocator,
     node_name,
     node_namespace,
-    _demangle_service_reply_from_topic,
-    _demangle_service_type_only,
+    rmw_gurumdds_cpp::demangle_service_reply_from_topic,
+    rmw_gurumdds_cpp::demangle_service_type_only,
     false,
-    __get_reader_names_and_types_by_node,
+    get_reader_names_and_types_by_node,
     service_names_and_types);
 }
 }  // extern "C"
