@@ -14,9 +14,6 @@
 
 #include <string>
 #include <sstream>
-#include <limits>
-#include <thread>
-#include <chrono>
 
 #include "rcutils/error_handling.h"
 #include "rcutils/types.h"
@@ -29,18 +26,25 @@
 #include "rmw/types.h"
 #include "rmw/validate_full_topic_name.h"
 
-#include "rmw_gurumdds_cpp/gid.hpp"
+#include "rmw_dds_common/qos.hpp"
+
 #include "rmw_gurumdds_cpp/graph_cache.hpp"
 #include "rmw_gurumdds_cpp/identifier.hpp"
-#include "rmw_gurumdds_cpp/namespace_prefix.hpp"
 #include "rmw_gurumdds_cpp/names_and_types_helpers.hpp"
+#include "rmw_gurumdds_cpp/namespace_prefix.hpp"
 #include "rmw_gurumdds_cpp/qos.hpp"
 #include "rmw_gurumdds_cpp/rmw_context_impl.hpp"
+#include "rmw_gurumdds_cpp/gid.hpp"
 #include "rmw_gurumdds_cpp/rmw_publisher.hpp"
-#include "rmw_gurumdds_cpp/types.hpp"
+#include "rmw_gurumdds_cpp/type_support.hpp"
+#include "rmw_gurumdds_cpp/type_support_common.hpp"
+#include "rmw_gurumdds_cpp/type_support_service.hpp"
+#include "rmw_gurumdds_cpp/event_info_common.hpp"
 
+namespace rmw_gurumdds_cpp
+{
 rmw_publisher_t *
-__rmw_create_publisher(
+create_publisher(
   rmw_context_impl_t * const ctx,
   const rmw_node_t * node,
   dds_DomainParticipant * const participant,
@@ -67,7 +71,7 @@ __rmw_create_publisher(
   }
 
   rmw_publisher_t * rmw_publisher = nullptr;
-  GurumddsPublisherInfo * publisher_info = nullptr;
+  PublisherInfo * publisher_info = nullptr;
   dds_DataWriter * topic_writer = nullptr;
   dds_DataWriterQos datawriter_qos;
   dds_Topic * topic = nullptr;
@@ -82,8 +86,8 @@ __rmw_create_publisher(
     return nullptr;
   }
 
-  std::string processed_topic_name = create_topic_name(
-    ros_topic_prefix, topic_name, "", qos_policies);
+  std::string processed_topic_name = rmw_gurumdds_cpp::create_topic_name(
+    rmw_gurumdds_cpp::ros_topic_prefix, topic_name, "", qos_policies);
 
   std::string metastring =
     create_metastring(type_support->data, type_support->typesupport_identifier);
@@ -92,15 +96,8 @@ __rmw_create_publisher(
     return nullptr;
   }
 
-  dds_typesupport = dds_TypeSupport_create(metastring.c_str());
+  dds_typesupport = create_type_support_and_register(participant, type_support, type_name, metastring);
   if (dds_typesupport == nullptr) {
-    RMW_SET_ERROR_MSG("failed to create typesupport");
-    return nullptr;
-  }
-
-  ret = dds_TypeSupport_register_type(dds_typesupport, participant, type_name.c_str());
-  if (ret != dds_RETCODE_OK) {
-    RMW_SET_ERROR_MSG("failed to register type to domain participant");
     return nullptr;
   }
 
@@ -127,6 +124,8 @@ __rmw_create_publisher(
       RMW_SET_ERROR_MSG("failed to finalize topic qos");
       return nullptr;
     }
+
+    TopicEventListener::associate_listener(topic);
   } else {
     dds_Duration_t timeout;
     timeout.sec = 0;
@@ -138,7 +137,8 @@ __rmw_create_publisher(
     }
   }
 
-  if (!get_datawriter_qos(pub, qos_policies, &datawriter_qos)) {
+  const rosidl_type_hash_t& type_hash = *type_support->get_type_hash_func(type_support);
+  if (!rmw_gurumdds_cpp::get_datawriter_qos(pub, qos_policies, type_hash, &datawriter_qos)) {
     // Error message already set
     return nullptr;
   }
@@ -156,19 +156,58 @@ __rmw_create_publisher(
     return nullptr;
   }
 
-  publisher_info = new(std::nothrow) GurumddsPublisherInfo();
+  publisher_info = new(std::nothrow) PublisherInfo();
   if (publisher_info == nullptr) {
-    RMW_SET_ERROR_MSG("failed to allocate GurumddsPublisherInfo");
+    RMW_SET_ERROR_MSG("failed to allocate PublisherInfo");
     return nullptr;
   }
 
+  dds_DataWriterListener listener{};
+  listener.on_offered_deadline_missed = [](const dds_DataWriter * topic_writer,
+                                           const dds_OfferedDeadlineMissedStatus * status) {
+    dds_DataWriter * writer = const_cast<dds_DataWriter *>(topic_writer);
+    PublisherInfo * info = static_cast<PublisherInfo*>(dds_DataWriter_get_listener_context(writer));
+    info->on_offered_deadline_missed(*status);
+  };
+
+  listener.on_offered_incompatible_qos = [](const dds_DataWriter * topic_writer,
+                                            const dds_OfferedIncompatibleQosStatus * status) {
+    dds_DataWriter * writer = const_cast<dds_DataWriter *>(topic_writer);
+    PublisherInfo * info = static_cast<PublisherInfo*>(dds_DataWriter_get_listener_context(writer));
+    info->on_offered_incompatible_qos(*status);
+  };
+
+  listener.on_liveliness_lost = [](const dds_DataWriter * topic_writer, const dds_LivelinessLostStatus * status) {
+    dds_DataWriter * writer = const_cast<dds_DataWriter *>(topic_writer);
+    PublisherInfo * info = static_cast<PublisherInfo*>(dds_DataWriter_get_listener_context(writer));
+    info->on_liveliness_lost(*status);
+  };
+
+  listener.on_publication_matched = [](const dds_DataWriter * topic_writer,
+                                       const dds_PublicationMatchedStatus * status) {
+    dds_DataWriter * writer = const_cast<dds_DataWriter *>(topic_writer);
+    PublisherInfo * info = static_cast<PublisherInfo*>(dds_DataWriter_get_listener_context(writer));
+    info->on_publication_matched(*status);
+  };
+
+  dds_DataWriter_set_listener_context(topic_writer, publisher_info);
   publisher_info->topic_writer = topic_writer;
+  publisher_info->topic_listener = listener;
   publisher_info->rosidl_message_typesupport = type_support;
   publisher_info->implementation_identifier = RMW_GURUMDDS_ID;
   publisher_info->sequence_number = 0;
   publisher_info->ctx = ctx;
+  publisher_info->event_guard_cond[RMW_EVENT_LIVELINESS_LOST] = dds_GuardCondition_create();
+  publisher_info->event_guard_cond[RMW_EVENT_OFFERED_DEADLINE_MISSED] = dds_GuardCondition_create();
+  publisher_info->event_guard_cond[RMW_EVENT_OFFERED_QOS_INCOMPATIBLE] = dds_GuardCondition_create();
+  publisher_info->event_guard_cond[RMW_EVENT_PUBLISHER_INCOMPATIBLE_TYPE] = dds_GuardCondition_create();
+  publisher_info->event_guard_cond[RMW_EVENT_PUBLICATION_MATCHED] = dds_GuardCondition_create();
+  dds_TypeSupport* reader_dds_type = dds_DataWriter_get_typesupport(topic_writer);
+  set_type_support_ops(reader_dds_type, type_support);
 
-  entity_get_gid(
+  TopicEventListener::add_event(topic, publisher_info);
+
+  rmw_gurumdds_cpp::entity_get_gid(
     reinterpret_cast<dds_Entity *>(publisher_info->topic_writer),
     publisher_info->publisher_gid);
 
@@ -195,7 +234,7 @@ __rmw_create_publisher(
     RCUTILS_LOG_ERROR_NAMED(RMW_GURUMDDS_ID, "failed to allocate publisher's topic name");
     return nullptr;
   }
-  memcpy(
+  std::memcpy(
     const_cast<char *>(rmw_publisher->topic_name),
     topic_name,
     strlen(topic_name) + 1);
@@ -203,7 +242,7 @@ __rmw_create_publisher(
   rmw_publisher->can_loan_messages = false;
 
   if (!internal) {
-    if (graph_on_publisher_created(ctx, node, publisher_info) != RMW_RET_OK) {
+    if (rmw_gurumdds_cpp::graph_cache::on_publisher_created(ctx, node, publisher_info) != RMW_RET_OK) {
       RCUTILS_LOG_ERROR_NAMED(RMW_GURUMDDS_ID, "failed to update graph for publisher");
       return nullptr;
     }
@@ -217,13 +256,13 @@ __rmw_create_publisher(
 }
 
 rmw_ret_t
-__rmw_destroy_publisher(
+destroy_publisher(
   rmw_context_impl_t * const ctx,
   rmw_publisher_t * const publisher)
 {
   std::lock_guard<std::mutex> guard(ctx->endpoint_mutex);
 
-  auto publisher_info = static_cast<GurumddsPublisherInfo *>(publisher->data);
+  auto publisher_info = static_cast<PublisherInfo *>(publisher->data);
   if (publisher_info == nullptr) {
     RMW_SET_ERROR_MSG("invalid publisher data");
     return RMW_RET_ERROR;
@@ -238,6 +277,7 @@ __rmw_destroy_publisher(
       return RMW_RET_ERROR;
     }
     publisher_info->topic_writer = nullptr;
+    TopicEventListener::remove_event(topic, publisher_info);
 
     ret = dds_DomainParticipant_delete_topic(ctx->participant, topic);
     if (ret == dds_RETCODE_PRECONDITION_NOT_MET) {
@@ -245,6 +285,14 @@ __rmw_destroy_publisher(
     } else if (ret != dds_RETCODE_OK) {
       RMW_SET_ERROR_MSG("failed to delete topic");
       return RMW_RET_ERROR;
+    } else {
+      TopicEventListener::disassociate_Listener(topic);
+    }
+  }
+
+  for(auto condition: publisher_info->event_guard_cond) {
+    if(nullptr != condition) {
+      dds_GuardCondition_delete(condition);
     }
   }
 
@@ -254,6 +302,87 @@ __rmw_destroy_publisher(
   return RMW_RET_OK;
 }
 
+rmw_ret_t publish(
+  const char* identifier,
+  const rmw_publisher_t* publisher,
+  const void* ros_message,
+  rmw_publisher_allocation_t* allocation) {
+  RCUTILS_UNUSED(allocation);
+  RMW_CHECK_ARGUMENT_FOR_NULL(publisher, RMW_RET_INVALID_ARGUMENT);
+  RMW_CHECK_ARGUMENT_FOR_NULL(ros_message, RMW_RET_INVALID_ARGUMENT);
+  RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
+    publisher,
+    identifier,
+    RMW_GURUMDDS_ID,
+    return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
+
+
+  auto publisher_info = static_cast<PublisherInfo *>(publisher->data);
+  RCUTILS_CHECK_ARGUMENT_FOR_NULL(publisher_info, RMW_RET_ERROR);
+
+  dds_DataWriter * topic_writer = publisher_info->topic_writer;
+  RCUTILS_CHECK_ARGUMENT_FOR_NULL(topic_writer, RMW_RET_ERROR);
+
+  const rosidl_message_type_support_t * rosidl_typesupport =
+    publisher_info->rosidl_message_typesupport;
+  if (rosidl_typesupport == nullptr) {
+    RMW_SET_ERROR_MSG("rosidl typesupport handle is null");
+    return RMW_RET_ERROR;
+  }
+
+  size_t size = 0;
+  void * dds_message = allocate_message(
+    rosidl_typesupport->data,
+    rosidl_typesupport->typesupport_identifier,
+    ros_message,
+    &size,
+    false
+  );
+
+  if (dds_message == nullptr) {
+    // Error message already set
+    return RMW_RET_ERROR;
+  }
+
+  bool result = serialize_ros_to_cdr(
+    rosidl_typesupport->data,
+    rosidl_typesupport->typesupport_identifier,
+    ros_message,
+    dds_message,
+    size
+  );
+  if (!result) {
+    RMW_SET_ERROR_MSG("failed to serialize message");
+    free(dds_message);
+    return RMW_RET_ERROR;
+  }
+
+  dds_SampleInfoEx sampleinfo_ex;
+  std::memset(&sampleinfo_ex, 0, sizeof(dds_SampleInfoEx));
+  ros_sn_to_dds_sn(++publisher_info->sequence_number, &sampleinfo_ex.seq);
+  rmw_gurumdds_cpp::ros_guid_to_dds_guid(
+      reinterpret_cast<const uint8_t *>(publisher_info->publisher_gid.data),
+      reinterpret_cast<uint8_t *>(&sampleinfo_ex.src_guid));
+
+  dds_ReturnCode_t ret = dds_DataWriter_raw_write_w_sampleinfoex(
+      topic_writer, dds_message, size, &sampleinfo_ex);
+
+  if (ret != dds_RETCODE_OK) {
+    std::stringstream errmsg;
+    errmsg << "failed to publish data: " << dds_ReturnCode_to_string(ret) << ", " << ret;
+    RMW_SET_ERROR_MSG(errmsg.str().c_str());
+    free(dds_message);
+    return RMW_RET_ERROR;
+  }
+
+  RCUTILS_LOG_DEBUG_NAMED(RMW_GURUMDDS_ID, "Published data on topic %s", publisher->topic_name);
+
+  free(dds_message);
+
+  return RMW_RET_OK;
+}
+} // namespace rmw_gurumdds_cpp
+
 extern "C"
 {
 rmw_ret_t
@@ -262,9 +391,9 @@ rmw_init_publisher_allocation(
   const rosidl_runtime_c__Sequence__bound * message_bounds,
   rmw_publisher_allocation_t * allocation)
 {
-  (void)type_support;
-  (void)message_bounds;
-  (void)allocation;
+  RCUTILS_UNUSED(type_support);
+  RCUTILS_UNUSED(message_bounds);
+  RCUTILS_UNUSED(allocation);
 
   RMW_SET_ERROR_MSG("rmw_init_publisher_allocation is not supported");
   return RMW_RET_UNSUPPORTED;
@@ -273,7 +402,7 @@ rmw_init_publisher_allocation(
 rmw_ret_t
 rmw_fini_publisher_allocation(rmw_publisher_allocation_t * allocation)
 {
-  (void)allocation;
+  RCUTILS_UNUSED(allocation);
 
   RMW_SET_ERROR_MSG("rmw_fini_publisher_allocation is not supported");
   return RMW_RET_UNSUPPORTED;
@@ -302,9 +431,17 @@ rmw_create_publisher(
   RMW_CHECK_ARGUMENT_FOR_NULL(qos_policies, nullptr);
   RMW_CHECK_ARGUMENT_FOR_NULL(publisher_options, nullptr);
 
-  if (!qos_policies->avoid_ros_namespace_conventions) {
+  // Adapt any 'best available' QoS options
+  rmw_qos_profile_t adapted_qos_policies = *qos_policies;
+  rmw_ret_t ret = rmw_dds_common::qos_profile_get_best_available_for_topic_publisher(
+    node, topic_name, &adapted_qos_policies, rmw_get_subscriptions_info_by_topic);
+  if (ret != RMW_RET_OK) {
+    return nullptr;
+  }
+
+  if (!adapted_qos_policies.avoid_ros_namespace_conventions) {
     int validation_result = RMW_TOPIC_VALID;
-    rmw_ret_t ret = rmw_validate_full_topic_name(topic_name, &validation_result, nullptr);
+    ret = rmw_validate_full_topic_name(topic_name, &validation_result, nullptr);
     if (ret != RMW_RET_OK) {
       return nullptr;
     }
@@ -326,14 +463,14 @@ rmw_create_publisher(
   rmw_context_impl_t * ctx = node->context->impl;
 
   rmw_publisher_t * const rmw_pub =
-    __rmw_create_publisher(
+    rmw_gurumdds_cpp::create_publisher(
     ctx,
     node,
     ctx->participant,
     ctx->publisher,
     type_supports,
     topic_name,
-    qos_policies,
+    &adapted_qos_policies,
     publisher_options,
     ctx->localhost_only);
 
@@ -364,7 +501,7 @@ rmw_publisher_count_matched_subscriptions(
     return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
   RMW_CHECK_ARGUMENT_FOR_NULL(subscription_count, RMW_RET_INVALID_ARGUMENT);
 
-  auto publisher_info = static_cast<GurumddsPublisherInfo *>(publisher->data);
+  auto publisher_info = static_cast<rmw_gurumdds_cpp::PublisherInfo *>(publisher->data);
   if (publisher_info == nullptr) {
     RMW_SET_ERROR_MSG("publisher internal data is invalid");
     return RMW_RET_ERROR;
@@ -394,7 +531,7 @@ rmw_publisher_assert_liveliness(const rmw_publisher_t * publisher)
 {
   RMW_CHECK_ARGUMENT_FOR_NULL(publisher, RMW_RET_INVALID_ARGUMENT);
 
-  auto publisher_info = static_cast<GurumddsPublisherInfo *>(publisher->data);
+  auto publisher_info = static_cast<rmw_gurumdds_cpp::PublisherInfo *>(publisher->data);
   if (publisher_info == nullptr) {
     RMW_SET_ERROR_MSG("publisher internal data is invalid");
     return RMW_RET_ERROR;
@@ -423,13 +560,13 @@ rmw_publisher_wait_for_all_acked(const rmw_publisher_t * publisher, rmw_time_t w
     RMW_GURUMDDS_ID,
     return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
 
-  auto publisher_info = static_cast<GurumddsPublisherInfo *>(publisher->data);
+  auto publisher_info = static_cast<rmw_gurumdds_cpp::PublisherInfo *>(publisher->data);
   if (publisher_info == nullptr) {
     RMW_SET_ERROR_MSG("publisher internal data is invalid");
     return RMW_RET_ERROR;
   }
 
-  dds_Duration_t timeout = rmw_time_to_dds(wait_timeout);
+  dds_Duration_t timeout = rmw_gurumdds_cpp::rmw_time_to_dds(wait_timeout);
   dds_ReturnCode_t ret = dds_DataWriter_wait_for_acknowledgments(
     publisher_info->topic_writer, &timeout);
 
@@ -461,14 +598,14 @@ rmw_destroy_publisher(rmw_node_t * node, rmw_publisher_t * publisher)
 
   rmw_context_impl_t * ctx = node->context->impl;
 
-  if (graph_on_publisher_deleted(
-      ctx, node, reinterpret_cast<GurumddsPublisherInfo *>(publisher->data)))
+  if (rmw_gurumdds_cpp::graph_cache::on_publisher_deleted(
+      ctx, node, reinterpret_cast<rmw_gurumdds_cpp::PublisherInfo *>(publisher->data)))
   {
     RCUTILS_LOG_ERROR_NAMED(RMW_GURUMDDS_ID, "failed to update graph for publisher");
     return RMW_RET_ERROR;
   }
 
-  rmw_ret_t ret = __rmw_destroy_publisher(ctx, publisher);
+  rmw_ret_t ret = rmw_gurumdds_cpp::destroy_publisher(ctx, publisher);
 
   if (ret == RMW_RET_OK) {
     if (publisher->topic_name != nullptr) {
@@ -497,7 +634,7 @@ rmw_get_gid_for_publisher(const rmw_publisher_t * publisher, rmw_gid_t * gid)
     RMW_GURUMDDS_ID,
     return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
 
-  auto publisher_info = static_cast<GurumddsPublisherInfo *>(publisher->data);
+  auto publisher_info = static_cast<rmw_gurumdds_cpp::PublisherInfo *>(publisher->data);
   if (publisher_info == nullptr) {
     RMW_SET_ERROR_MSG("publisher info handle is null");
     return RMW_RET_ERROR;
@@ -521,7 +658,7 @@ rmw_publisher_get_actual_qos(
     RMW_GURUMDDS_ID,
     return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
 
-  auto publisher_info = static_cast<GurumddsPublisherInfo *>(publisher->data);
+  auto publisher_info = static_cast<rmw_gurumdds_cpp::PublisherInfo *>(publisher->data);
   if (publisher_info == nullptr) {
     RMW_SET_ERROR_MSG("publisher internal data is invalid");
     return RMW_RET_ERROR;
@@ -540,13 +677,13 @@ rmw_publisher_get_actual_qos(
     return RMW_RET_ERROR;
   }
 
-  qos->reliability = convert_reliability(&dds_qos.reliability);
-  qos->durability = convert_durability(&dds_qos.durability);
-  qos->deadline = convert_deadline(&dds_qos.deadline);
-  qos->lifespan = convert_lifespan(&dds_qos.lifespan);
-  qos->liveliness = convert_liveliness(&dds_qos.liveliness);
-  qos->liveliness_lease_duration = convert_liveliness_lease_duration(&dds_qos.liveliness);
-  qos->history = convert_history(&dds_qos.history);
+  qos->reliability = rmw_gurumdds_cpp::convert_reliability(&dds_qos.reliability);
+  qos->durability = rmw_gurumdds_cpp::convert_durability(&dds_qos.durability);
+  qos->deadline = rmw_gurumdds_cpp::convert_deadline(&dds_qos.deadline);
+  qos->lifespan = rmw_gurumdds_cpp::convert_lifespan(&dds_qos.lifespan);
+  qos->liveliness = rmw_gurumdds_cpp::convert_liveliness(&dds_qos.liveliness);
+  qos->liveliness_lease_duration = rmw_gurumdds_cpp::convert_liveliness_lease_duration(&dds_qos.liveliness);
+  qos->history = rmw_gurumdds_cpp::convert_history(&dds_qos.history);
   qos->depth = static_cast<size_t>(dds_qos.history.depth);
 
   ret = dds_DataWriterQos_finalize(&dds_qos);
@@ -564,88 +701,8 @@ rmw_publish(
   const void * ros_message,
   rmw_publisher_allocation_t * allocation)
 {
-  (void)allocation;
   RMW_CHECK_ARGUMENT_FOR_NULL(publisher, RMW_RET_INVALID_ARGUMENT);
-  RMW_CHECK_ARGUMENT_FOR_NULL(ros_message, RMW_RET_INVALID_ARGUMENT);
-  RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
-    publisher,
-    publisher->implementation_identifier,
-    RMW_GURUMDDS_ID,
-    return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
-
-  auto publisher_info = static_cast<GurumddsPublisherInfo *>(publisher->data);
-  RCUTILS_CHECK_ARGUMENT_FOR_NULL(publisher_info, RMW_RET_ERROR);
-
-  dds_DataWriter * topic_writer = publisher_info->topic_writer;
-  RCUTILS_CHECK_ARGUMENT_FOR_NULL(topic_writer, RMW_RET_ERROR);
-
-  const rosidl_message_type_support_t * rosidl_typesupport =
-    publisher_info->rosidl_message_typesupport;
-  if (rosidl_typesupport == nullptr) {
-    RMW_SET_ERROR_MSG("rosidl typesupport handle is null");
-    return RMW_RET_ERROR;
-  }
-
-  size_t size = 0;
-  void * dds_message = allocate_message(
-    rosidl_typesupport->data,
-    rosidl_typesupport->typesupport_identifier,
-    ros_message,
-    &size,
-    false
-  );
-  if (dds_message == nullptr) {
-    // Error message already set
-    return RMW_RET_ERROR;
-  }
-
-  bool result = serialize_ros_to_cdr(
-    rosidl_typesupport->data,
-    rosidl_typesupport->typesupport_identifier,
-    ros_message,
-    dds_message,
-    size
-  );
-  if (!result) {
-    RMW_SET_ERROR_MSG("failed to serialize message");
-    free(dds_message);
-    return RMW_RET_ERROR;
-  }
-
-  dds_SampleInfoEx sampleinfo_ex;
-  memset(&sampleinfo_ex, 0, sizeof(dds_SampleInfoEx));
-  ros_sn_to_dds_sn(++publisher_info->sequence_number, &sampleinfo_ex.seq);
-  ros_guid_to_dds_guid(
-    reinterpret_cast<int8_t *>(publisher_info->publisher_gid.data),
-    reinterpret_cast<int8_t *>(&sampleinfo_ex.src_guid));
-
-  dds_ReturnCode_t ret = dds_DataWriter_raw_write_w_sampleinfoex(
-    topic_writer, dds_message, size, &sampleinfo_ex);
-
-  const char * errstr;
-  if (ret == dds_RETCODE_OK) {
-    errstr = "dds_RETCODE_OK";
-  } else if (ret == dds_RETCODE_TIMEOUT) {
-    errstr = "dds_RETCODE_TIMEOUT";
-  } else if (ret == dds_RETCODE_OUT_OF_RESOURCES) {
-    errstr = "dds_RETCODE_OUT_OF_RESOURCES";
-  } else {
-    errstr = "dds_RETCODE_ERROR";
-  }
-
-  if (ret != dds_RETCODE_OK) {
-    std::stringstream errmsg;
-    errmsg << "failed to publish data: " << errstr << ", " << ret;
-    RMW_SET_ERROR_MSG(errmsg.str().c_str());
-    free(dds_message);
-    return RMW_RET_ERROR;
-  }
-
-  RCUTILS_LOG_DEBUG_NAMED(RMW_GURUMDDS_ID, "Published data on topic %s", publisher->topic_name);
-
-  free(dds_message);
-
-  return RMW_RET_OK;
+  return rmw_gurumdds_cpp::publish(publisher->implementation_identifier, publisher, ros_message, allocation);
 }
 
 rmw_ret_t
@@ -654,7 +711,7 @@ rmw_publish_serialized_message(
   const rmw_serialized_message_t * serialized_message,
   rmw_publisher_allocation_t * allocation)
 {
-  (void)allocation;
+  RCUTILS_UNUSED(allocation);
   RMW_CHECK_ARGUMENT_FOR_NULL(publisher, RMW_RET_INVALID_ARGUMENT);
   RMW_CHECK_ARGUMENT_FOR_NULL(serialized_message, RMW_RET_INVALID_ARGUMENT);
   RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
@@ -663,18 +720,17 @@ rmw_publish_serialized_message(
     RMW_GURUMDDS_ID,
     return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
 
-  auto publisher_info = static_cast<GurumddsPublisherInfo *>(publisher->data);
+  auto publisher_info = static_cast<rmw_gurumdds_cpp::PublisherInfo *>(publisher->data);
   RCUTILS_CHECK_ARGUMENT_FOR_NULL(publisher_info, RMW_RET_ERROR);
 
   dds_DataWriter * topic_writer = publisher_info->topic_writer;
   RCUTILS_CHECK_ARGUMENT_FOR_NULL(topic_writer, RMW_RET_ERROR);
 
-  dds_SampleInfoEx sampleinfo_ex;
-  memset(&sampleinfo_ex, 0, sizeof(dds_SampleInfoEx));
-  ros_sn_to_dds_sn(++publisher_info->sequence_number, &sampleinfo_ex.seq);
-  ros_guid_to_dds_guid(
-    reinterpret_cast<int8_t *>(publisher_info->publisher_gid.data),
-    reinterpret_cast<int8_t *>(&sampleinfo_ex.src_guid));
+  dds_SampleInfoEx sampleinfo_ex{};
+  rmw_gurumdds_cpp::ros_sn_to_dds_sn(++publisher_info->sequence_number, &sampleinfo_ex.seq);
+  rmw_gurumdds_cpp::ros_guid_to_dds_guid(
+    reinterpret_cast<const uint8_t *>(publisher_info->publisher_gid.data),
+    reinterpret_cast<uint8_t *>(&sampleinfo_ex.src_guid));
 
   dds_ReturnCode_t ret = dds_DataWriter_raw_write_w_sampleinfoex(
     topic_writer,
@@ -683,20 +739,9 @@ rmw_publish_serialized_message(
     &sampleinfo_ex
   );
 
-  const char * errstr;
-  if (ret == dds_RETCODE_OK) {
-    errstr = "dds_RETCODE_OK";
-  } else if (ret == dds_RETCODE_TIMEOUT) {
-    errstr = "dds_RETCODE_TIMEOUT";
-  } else if (ret == dds_RETCODE_OUT_OF_RESOURCES) {
-    errstr = "dds_RETCODE_OUT_OF_RESOURCES";
-  } else {
-    errstr = "dds_RETCODE_ERROR";
-  }
-
   if (ret != dds_RETCODE_OK) {
     std::stringstream errmsg;
-    errmsg << "failed to publish data: " << errstr << ", " << ret;
+    errmsg << "failed to publish data: " << dds_ReturnCode_to_string(ret) << ", " << ret;
     RMW_SET_ERROR_MSG(errmsg.str().c_str());
     return RMW_RET_ERROR;
   }
@@ -712,9 +757,9 @@ rmw_publish_loaned_message(
   void * ros_message,
   rmw_publisher_allocation_t * allocation)
 {
-  (void)publisher;
-  (void)ros_message;
-  (void)allocation;
+  RCUTILS_UNUSED(publisher);
+  RCUTILS_UNUSED(ros_message);
+  RCUTILS_UNUSED(allocation);
 
   RMW_SET_ERROR_MSG("rmw_publish_loaned_message is not supported");
   return RMW_RET_UNSUPPORTED;
@@ -726,9 +771,9 @@ rmw_borrow_loaned_message(
   const rosidl_message_type_support_t * type_support,
   void ** ros_message)
 {
-  (void)publisher;
-  (void)type_support;
-  (void)ros_message;
+  RCUTILS_UNUSED(publisher);
+  RCUTILS_UNUSED(type_support);
+  RCUTILS_UNUSED(ros_message);
 
   RMW_SET_ERROR_MSG("rmw_borrow_loaned_message is not supported");
   return RMW_RET_UNSUPPORTED;
@@ -739,8 +784,8 @@ rmw_return_loaned_message_from_publisher(
   const rmw_publisher_t * publisher,
   void * loaned_message)
 {
-  (void)publisher;
-  (void)loaned_message;
+  RCUTILS_UNUSED(publisher);
+  RCUTILS_UNUSED(loaned_message);
 
   RMW_SET_ERROR_MSG("rmw_return_loaned_message_from_publisher is not supported");
   return RMW_RET_UNSUPPORTED;
