@@ -14,28 +14,29 @@
 
 #include <chrono>
 #include <string>
-#include <thread>
 #include <utility>
 
-#include "rcutils/logging_macros.h"
 #include "rcutils/error_handling.h"
+#include "rcutils/logging_macros.h"
 
 #include "rmw/allocators.h"
-#include "rmw/get_service_names_and_types.h"
-#include "rmw/names_and_types.h"
 #include "rmw/rmw.h"
 #include "rmw/validate_full_topic_name.h"
 
-#include "rmw_gurumdds_cpp/gid.hpp"
+#include "rmw_dds_common/qos.hpp"
+
+#include "rmw_gurumdds_cpp/event_converter.hpp"
 #include "rmw_gurumdds_cpp/graph_cache.hpp"
 #include "rmw_gurumdds_cpp/identifier.hpp"
 #include "rmw_gurumdds_cpp/names_and_types_helpers.hpp"
 #include "rmw_gurumdds_cpp/namespace_prefix.hpp"
 #include "rmw_gurumdds_cpp/qos.hpp"
 #include "rmw_gurumdds_cpp/rmw_context_impl.hpp"
-#include "rmw_gurumdds_cpp/types.hpp"
+#include "rmw_gurumdds_cpp/gid.hpp"
+#include "rmw_gurumdds_cpp/event_info_common.hpp"
+#include "rmw_gurumdds_cpp/event_info_service.hpp"
 
-#include "type_support_service.hpp"
+#include "rmw_gurumdds_cpp/type_support_service.hpp"
 
 extern "C"
 {
@@ -51,7 +52,7 @@ rmw_create_service(
     node,
     node->implementation_identifier,
     RMW_GURUMDDS_ID,
-    return nullptr)
+    return nullptr);
   RMW_CHECK_ARGUMENT_FOR_NULL(type_supports, nullptr);
   RMW_CHECK_ARGUMENT_FOR_NULL(service_name, nullptr);
   if (strlen(service_name) == 0) {
@@ -60,7 +61,11 @@ rmw_create_service(
   }
   RMW_CHECK_ARGUMENT_FOR_NULL(qos_policies, nullptr);
 
-  if (!qos_policies->avoid_ros_namespace_conventions) {
+  // Adapt any 'best available' QoS options
+  rmw_qos_profile_t adapted_qos_policies =
+    rmw_dds_common::qos_profile_update_best_available_for_services(*qos_policies);
+
+  if (!adapted_qos_policies.avoid_ros_namespace_conventions) {
     int validation_result = RMW_TOPIC_VALID;
     rmw_ret_t ret = rmw_validate_full_topic_name(service_name, &validation_result, nullptr);
     if (ret != RMW_RET_OK) {
@@ -89,7 +94,7 @@ rmw_create_service(
   rmw_context_impl_t * ctx = node->context->impl;
   std::lock_guard<std::mutex> guard(ctx->endpoint_mutex);
 
-  GurumddsServiceInfo * service_info = nullptr;
+  rmw_gurumdds_cpp::ServiceInfo * service_info = nullptr;
   rmw_service_t * rmw_service = nullptr;
 
   dds_DomainParticipant * participant = ctx->participant;
@@ -100,6 +105,10 @@ rmw_create_service(
   dds_DataWriterQos datawriter_qos;
 
   dds_DataReader * request_reader = nullptr;
+  dds_DataReaderListener request_listener;
+  dds_DataSeq* data_seq = nullptr;
+  dds_SampleInfoSeq* info_seq = nullptr;
+  dds_UnsignedLongSeq* raw_data_sizes = nullptr;
   dds_DataWriter * response_writer = nullptr;
   dds_ReadCondition * read_condition = nullptr;
   dds_TypeSupport * request_typesupport = nullptr;
@@ -120,9 +129,11 @@ rmw_create_service(
   std::string request_metastring;
   std::string response_metastring;
 
+  const rosidl_type_hash_t* type_hash;
+
   // Create topic and type name strings
   service_type_name =
-    create_service_type_name(type_support->data, type_support->typesupport_identifier);
+    rmw_gurumdds_cpp::create_service_type_name(type_support->data, type_support->typesupport_identifier);
   request_type_name = service_type_name.first;
   response_type_name = service_type_name.second;
   if (request_type_name.empty() || response_type_name.empty()) {
@@ -132,30 +143,19 @@ rmw_create_service(
 
   request_topic_name.reserve(256);
   response_topic_name.reserve(256);
-  request_topic_name = create_topic_name(
-    ros_service_requester_prefix, service_name, "Request", qos_policies);
-  response_topic_name = create_topic_name(
-    ros_service_response_prefix, service_name, "Reply", qos_policies);
+  request_topic_name = rmw_gurumdds_cpp::create_topic_name(
+    rmw_gurumdds_cpp::ros_service_requester_prefix, service_name, "Request", &adapted_qos_policies);
+  response_topic_name = rmw_gurumdds_cpp::create_topic_name(
+    rmw_gurumdds_cpp::ros_service_response_prefix, service_name, "Reply", &adapted_qos_policies);
 
   service_metastring =
-    create_service_metastring(type_support->data, type_support->typesupport_identifier);
+    rmw_gurumdds_cpp::create_service_metastring(type_support->data, type_support->typesupport_identifier);
   request_metastring = service_metastring.first;
   response_metastring = service_metastring.second;
   if (request_metastring.empty() || response_metastring.empty()) {
     RMW_SET_ERROR_MSG("failed to create metastring");
     return nullptr;
   }
-
-  // Set infos for this service(server)
-  service_info = new(std::nothrow) GurumddsServiceInfo();
-  if (service_info == nullptr) {
-    RMW_SET_ERROR_MSG("failed to allocate GurumddsServiceInfo");
-    goto fail;
-  }
-
-  service_info->implementation_identifier = RMW_GURUMDDS_ID;
-  service_info->service_typesupport = type_support;
-  service_info->ctx = ctx;
 
   request_typesupport = dds_TypeSupport_create(request_metastring.c_str());
   if (request_typesupport == nullptr) {
@@ -257,7 +257,8 @@ rmw_create_service(
     }
   }
 
-  if (!get_datareader_qos(subscriber, qos_policies, &datareader_qos)) {
+  type_hash = type_support->request_typesupport->get_type_hash_func(type_support->request_typesupport);
+  if (!rmw_gurumdds_cpp::get_datareader_qos(subscriber, &adapted_qos_policies, *type_hash, &datareader_qos)) {
     // Error message already set
     goto fail;
   }
@@ -269,7 +270,6 @@ rmw_create_service(
     dds_DataReaderQos_finalize(&datareader_qos);
     goto fail;
   }
-  service_info->request_reader = request_reader;
 
   ret = dds_DataReaderQos_finalize(&datareader_qos);
   if (ret != dds_RETCODE_OK) {
@@ -283,9 +283,9 @@ rmw_create_service(
     RMW_SET_ERROR_MSG("failed to create read condition");
     goto fail;
   }
-  service_info->read_condition = read_condition;
 
-  if (!get_datawriter_qos(publisher, qos_policies, &datawriter_qos)) {
+  type_hash = type_support->response_typesupport->get_type_hash_func(type_support->response_typesupport);
+  if (!rmw_gurumdds_cpp::get_datawriter_qos(publisher, &adapted_qos_policies, *type_hash, &datawriter_qos)) {
     // Error message already set
     goto fail;
   }
@@ -297,7 +297,6 @@ rmw_create_service(
     dds_DataWriterQos_finalize(&datawriter_qos);
     goto fail;
   }
-  service_info->response_writer = response_writer;
 
   ret = dds_DataWriterQos_finalize(&datawriter_qos);
   if (ret != dds_RETCODE_OK) {
@@ -305,10 +304,53 @@ rmw_create_service(
     goto fail;
   }
 
-  entity_get_gid(
+  service_info = new(std::nothrow) rmw_gurumdds_cpp::ServiceInfo();
+  if (service_info == nullptr) {
+    RMW_SET_ERROR_MSG("failed to allocate ServiceInfo");
+    goto fail;
+  }
+
+  data_seq = dds_DataSeq_create(1);
+  if (nullptr == data_seq) {
+    RMW_SET_ERROR_MSG("failed to allocate data_seq");
+    return nullptr;
+  }
+  info_seq = dds_SampleInfoSeq_create(1);
+  if (nullptr == info_seq) {
+    RMW_SET_ERROR_MSG("failed to allocate info_seq");
+    return nullptr;
+  }
+  raw_data_sizes = dds_UnsignedLongSeq_create(1);
+  if (nullptr == raw_data_sizes) {
+    RMW_SET_ERROR_MSG("failed to allocate raw_data_sizes");
+    return nullptr;
+  }
+
+  dds_DataReader_set_listener_context(request_reader, service_info);
+  request_listener.on_data_available = [](const dds_DataReader * request_reader){
+    dds_DataReader* reader = const_cast<dds_DataReader*>(request_reader);
+    rmw_gurumdds_cpp::ServiceInfo* info = static_cast<rmw_gurumdds_cpp::ServiceInfo*>(dds_DataReader_get_listener_context(reader));
+    std::lock_guard<std::mutex> guard(info->event_callback_data.mutex);
+    if(info->event_callback_data.callback) {
+      info->event_callback_data.callback(info->event_callback_data.user_data, info->count_unread());
+    }
+  };
+
+  service_info->response_writer = response_writer;
+  service_info->request_reader = request_reader;
+  service_info->read_condition = read_condition;
+  service_info->request_listener = request_listener;
+  service_info->data_seq = data_seq;
+  service_info->info_seq = info_seq;
+  service_info->raw_data_sizes = raw_data_sizes;
+  service_info->implementation_identifier = RMW_GURUMDDS_ID;
+  service_info->service_typesupport = type_support;
+  service_info->ctx = ctx;
+
+  rmw_gurumdds_cpp::entity_get_gid(
     reinterpret_cast<dds_Entity *>(service_info->response_writer),
     service_info->publisher_gid);
-  entity_get_gid(
+  rmw_gurumdds_cpp::entity_get_gid(
     reinterpret_cast<dds_Entity *>(service_info->request_reader),
     service_info->subscriber_gid);
 
@@ -317,7 +359,7 @@ rmw_create_service(
     RMW_SET_ERROR_MSG("failed to allocate memory for service");
     goto fail;
   }
-  memset(rmw_service, 0, sizeof(rmw_service_t));
+  std::memset(rmw_service, 0, sizeof(rmw_service_t));
   rmw_service->implementation_identifier = RMW_GURUMDDS_ID;
   rmw_service->data = service_info;
   rmw_service->service_name =
@@ -326,9 +368,9 @@ rmw_create_service(
     RMW_SET_ERROR_MSG("failed to allocate memory for service name");
     goto fail;
   }
-  memcpy(const_cast<char *>(rmw_service->service_name), service_name, strlen(service_name) + 1);
+  std::memcpy(const_cast<char *>(rmw_service->service_name), service_name, strlen(service_name) + 1);
 
-  if (graph_on_service_created(ctx, node, service_info) != RMW_RET_OK) {
+  if (rmw_gurumdds_cpp::graph_cache::on_service_created(ctx, node, service_info) != RMW_RET_OK) {
     RCUTILS_LOG_ERROR_NAMED(RMW_GURUMDDS_ID, "failed to update graph for service creation");
     goto fail;
   }
@@ -381,9 +423,8 @@ fail:
     dds_TypeSupport_delete(response_typesupport);
   }
 
-  if (service_info != nullptr) {
-    delete service_info;
-  }
+  delete service_info;
+
   return nullptr;
 }
 
@@ -407,7 +448,7 @@ rmw_destroy_service(rmw_node_t * node, rmw_service_t * service)
   rmw_context_impl_t * ctx = node->context->impl;
   std::lock_guard<std::mutex> guard(ctx->endpoint_mutex);
 
-  GurumddsServiceInfo * service_info = static_cast<GurumddsServiceInfo *>(service->data);
+  rmw_gurumdds_cpp::ServiceInfo * service_info = static_cast<rmw_gurumdds_cpp::ServiceInfo *>(service->data);
   if (service_info != nullptr) {
     if (service_info->response_writer != nullptr) {
       ret = dds_Publisher_delete_datawriter(
@@ -417,6 +458,10 @@ rmw_destroy_service(rmw_node_t * node, rmw_service_t * service)
         return RMW_RET_ERROR;
       }
     }
+
+    dds_DataSeq_delete(service_info->data_seq);
+    dds_SampleInfoSeq_delete(service_info->info_seq);
+    dds_UnsignedLongSeq_delete(service_info->raw_data_sizes);
 
     if (service_info->request_reader != nullptr) {
       if (service_info->read_condition != nullptr) {
@@ -435,7 +480,7 @@ rmw_destroy_service(rmw_node_t * node, rmw_service_t * service)
       }
     }
 
-    if (graph_on_service_deleted(ctx, node, service_info) != RMW_RET_OK) {
+    if (rmw_gurumdds_cpp::graph_cache::on_service_deleted(ctx, node, service_info) != RMW_RET_OK) {
       RCUTILS_LOG_ERROR_NAMED(RMW_GURUMDDS_ID, "failed to update graph for service deletion");
       return RMW_RET_ERROR;
     }
@@ -470,7 +515,7 @@ rmw_service_response_publisher_get_actual_qos(
     return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
   RMW_CHECK_ARGUMENT_FOR_NULL(qos, RMW_RET_INVALID_ARGUMENT);
 
-  GurumddsServiceInfo * service_info = static_cast<GurumddsServiceInfo *>(service->data);
+  rmw_gurumdds_cpp::ServiceInfo * service_info = static_cast<rmw_gurumdds_cpp::ServiceInfo *>(service->data);
   if (service_info == nullptr) {
     RMW_SET_ERROR_MSG("service info is null");
     return RMW_RET_ERROR;
@@ -489,13 +534,13 @@ rmw_service_response_publisher_get_actual_qos(
     return RMW_RET_ERROR;
   }
 
-  qos->reliability = convert_reliability(&dds_qos.reliability);
-  qos->durability = convert_durability(&dds_qos.durability);
-  qos->deadline = convert_deadline(&dds_qos.deadline);
-  qos->lifespan = convert_lifespan(&dds_qos.lifespan);
-  qos->liveliness = convert_liveliness(&dds_qos.liveliness);
-  qos->liveliness_lease_duration = convert_liveliness_lease_duration(&dds_qos.liveliness);
-  qos->history = convert_history(&dds_qos.history);
+  qos->reliability = rmw_gurumdds_cpp::convert_reliability(&dds_qos.reliability);
+  qos->durability = rmw_gurumdds_cpp::convert_durability(&dds_qos.durability);
+  qos->deadline = rmw_gurumdds_cpp::convert_deadline(&dds_qos.deadline);
+  qos->lifespan = rmw_gurumdds_cpp::convert_lifespan(&dds_qos.lifespan);
+  qos->liveliness = rmw_gurumdds_cpp::convert_liveliness(&dds_qos.liveliness);
+  qos->liveliness_lease_duration = rmw_gurumdds_cpp::convert_liveliness_lease_duration(&dds_qos.liveliness);
+  qos->history = rmw_gurumdds_cpp::convert_history(&dds_qos.history);
   qos->depth = static_cast<size_t>(dds_qos.history.depth);
 
   ret = dds_DataWriterQos_finalize(&dds_qos);
@@ -520,7 +565,7 @@ rmw_service_request_subscription_get_actual_qos(
     return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
   RMW_CHECK_ARGUMENT_FOR_NULL(qos, RMW_RET_INVALID_ARGUMENT);
 
-  GurumddsServiceInfo * service_info = static_cast<GurumddsServiceInfo *>(service->data);
+  rmw_gurumdds_cpp::ServiceInfo * service_info = static_cast<rmw_gurumdds_cpp::ServiceInfo *>(service->data);
   if (service_info == nullptr) {
     RMW_SET_ERROR_MSG("service info is null");
     return RMW_RET_ERROR;
@@ -539,12 +584,12 @@ rmw_service_request_subscription_get_actual_qos(
     return RMW_RET_ERROR;
   }
 
-  qos->reliability = convert_reliability(&dds_qos.reliability);
-  qos->durability = convert_durability(&dds_qos.durability);
-  qos->deadline = convert_deadline(&dds_qos.deadline);
-  qos->liveliness = convert_liveliness(&dds_qos.liveliness);
-  qos->liveliness_lease_duration = convert_liveliness_lease_duration(&dds_qos.liveliness);
-  qos->history = convert_history(&dds_qos.history);
+  qos->reliability = rmw_gurumdds_cpp::convert_reliability(&dds_qos.reliability);
+  qos->durability = rmw_gurumdds_cpp::convert_durability(&dds_qos.durability);
+  qos->deadline = rmw_gurumdds_cpp::convert_deadline(&dds_qos.deadline);
+  qos->liveliness = rmw_gurumdds_cpp::convert_liveliness(&dds_qos.liveliness);
+  qos->liveliness_lease_duration = rmw_gurumdds_cpp::convert_liveliness_lease_duration(&dds_qos.liveliness);
+  qos->history = rmw_gurumdds_cpp::convert_history(&dds_qos.history);
   qos->depth = static_cast<size_t>(dds_qos.history.depth);
 
   ret = dds_DataReaderQos_finalize(&dds_qos);
@@ -574,7 +619,7 @@ rmw_take_request(
 
   *taken = false;
 
-  GurumddsServiceInfo * service_info = static_cast<GurumddsServiceInfo *>(service->data);
+  rmw_gurumdds_cpp::ServiceInfo * service_info = static_cast<rmw_gurumdds_cpp::ServiceInfo *>(service->data);
   if (service_info == nullptr) {
     RMW_SET_ERROR_MSG("service info handle is null");
     return RMW_RET_ERROR;
@@ -649,8 +694,9 @@ rmw_take_request(
       int32_t sn_high = 0;
       uint32_t sn_low = 0;
       int8_t client_guid[16] = {0};
+      dds_SampleInfoEx * sampleinfo_ex = reinterpret_cast<dds_SampleInfoEx *>(sample_info);
 
-      bool res = deserialize_request_basic(
+      bool res = rmw_gurumdds_cpp::deserialize_request_basic(
         type_support->data,
         type_support->typesupport_identifier,
         ros_request,
@@ -673,10 +719,11 @@ rmw_take_request(
       request_header->source_timestamp =
         sample_info->source_timestamp.sec * static_cast<int64_t>(1000000000) +
         sample_info->source_timestamp.nanosec;
-      // TODO(clemjh): SampleInfo doesn't contain received_timestamp
-      request_header->received_timestamp = 0;
+      request_header->received_timestamp =
+        sampleinfo_ex->reception_timestamp.sec * static_cast<int64_t>(1000000000) +
+        sampleinfo_ex->reception_timestamp.nanosec;
       request_header->request_id.sequence_number = ((int64_t)sn_high) << 32 | sn_low;
-      memcpy(request_header->request_id.writer_guid, client_guid, 16);
+      std::memcpy(request_header->request_id.writer_guid, client_guid, RMW_GID_STORAGE_SIZE);
     }
 
     dds_DataReader_raw_return_loan(request_reader, data_values, sample_infos, sample_sizes);
@@ -719,10 +766,10 @@ rmw_take_request(
       int64_t sequence_number = 0;
       int8_t client_guid[16] = {0};
       dds_SampleInfoEx * sampleinfo_ex = reinterpret_cast<dds_SampleInfoEx *>(sample_info);
-      dds_guid_to_ros_guid(reinterpret_cast<int8_t *>(&sampleinfo_ex->src_guid), client_guid);
-      dds_sn_to_ros_sn(sampleinfo_ex->seq, &sequence_number);
+      rmw_gurumdds_cpp::dds_guid_to_ros_guid(reinterpret_cast<int8_t *>(&sampleinfo_ex->src_guid), client_guid);
+      rmw_gurumdds_cpp::dds_sn_to_ros_sn(sampleinfo_ex->seq, &sequence_number);
 
-      bool res = deserialize_request_enhanced(
+      bool res = rmw_gurumdds_cpp::deserialize_request_enhanced(
         type_support->data,
         type_support->typesupport_identifier,
         ros_request,
@@ -742,10 +789,11 @@ rmw_take_request(
       request_header->source_timestamp =
         sample_info->source_timestamp.sec * static_cast<int64_t>(1000000000) +
         sample_info->source_timestamp.nanosec;
-      // TODO(clemjh): SampleInfo doesn't contain received_timestamp
-      request_header->received_timestamp = 0;
+      request_header->received_timestamp =
+        sampleinfo_ex->reception_timestamp.sec * static_cast<int64_t>(1000000000) +
+        sampleinfo_ex->reception_timestamp.nanosec;
       request_header->request_id.sequence_number = sequence_number;
-      memcpy(request_header->request_id.writer_guid, client_guid, 16);
+      std::memcpy(request_header->request_id.writer_guid, client_guid, RMW_GID_STORAGE_SIZE);
     }
 
     dds_DataReader_raw_return_loan(request_reader, data_values, sample_infos, sample_sizes);
@@ -772,7 +820,7 @@ rmw_send_response(
   RMW_CHECK_ARGUMENT_FOR_NULL(request_header, RMW_RET_INVALID_ARGUMENT);
   RMW_CHECK_ARGUMENT_FOR_NULL(ros_response, RMW_RET_INVALID_ARGUMENT);
 
-  GurumddsServiceInfo * service_info = static_cast<GurumddsServiceInfo *>(service->data);
+  rmw_gurumdds_cpp::ServiceInfo * service_info = static_cast<rmw_gurumdds_cpp::ServiceInfo *>(service->data);
   if (service_info == nullptr) {
     RMW_SET_ERROR_MSG("service info handle is null");
     return RMW_RET_ERROR;
@@ -793,7 +841,7 @@ rmw_send_response(
   size_t size = 0;
 
   if (service_info->ctx->service_mapping_basic) {
-    void * dds_response = allocate_response_basic(
+    void * dds_response = rmw_gurumdds_cpp::allocate_response_basic(
       type_support->data,
       type_support->typesupport_identifier,
       ros_response,
@@ -804,7 +852,7 @@ rmw_send_response(
       return RMW_RET_ERROR;
     }
 
-    bool res = serialize_response_basic(
+    bool res = rmw_gurumdds_cpp::serialize_response_basic(
       type_support->data,
       type_support->typesupport_identifier,
       ros_response,
@@ -827,7 +875,7 @@ rmw_send_response(
     }
     free(dds_response);
   } else {
-    void * dds_response = allocate_response_enhanced(
+    void * dds_response = rmw_gurumdds_cpp::allocate_response_enhanced(
       type_support->data,
       type_support->typesupport_identifier,
       ros_response,
@@ -839,7 +887,7 @@ rmw_send_response(
       return RMW_RET_ERROR;
     }
 
-    bool res = serialize_response_enhanced(
+    bool res = rmw_gurumdds_cpp::serialize_response_enhanced(
       type_support->data,
       type_support->typesupport_identifier,
       ros_response,
@@ -854,11 +902,11 @@ rmw_send_response(
     }
 
     dds_SampleInfoEx sampleinfo_ex;
-    memset(&sampleinfo_ex, 0, sizeof(dds_SampleInfoEx));
-    ros_sn_to_dds_sn(request_header->sequence_number, &sampleinfo_ex.seq);
-    ros_guid_to_dds_guid(
+    std::memset(&sampleinfo_ex, 0, sizeof(dds_SampleInfoEx));
+    rmw_gurumdds_cpp::ros_sn_to_dds_sn(request_header->sequence_number, &sampleinfo_ex.seq);
+    rmw_gurumdds_cpp::ros_guid_to_dds_guid(
       request_header->writer_guid,
-      reinterpret_cast<int8_t *>(&sampleinfo_ex.src_guid));
+      reinterpret_cast<uint8_t *>(&sampleinfo_ex.src_guid));
 
     if (dds_DataWriter_raw_write_w_sampleinfoex(
         response_writer, dds_response, size, &sampleinfo_ex) != dds_RETCODE_OK)
@@ -879,11 +927,39 @@ rmw_service_set_on_new_request_callback(
   rmw_event_callback_t callback,
   const void * user_data)
 {
-  (void)rmw_service;
-  (void)callback;
-  (void)user_data;
+  RMW_CHECK_ARGUMENT_FOR_NULL(rmw_service, RMW_RET_INVALID_ARGUMENT);
+  RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
+    rmw_service,
+    rmw_service->implementation_identifier,
+    RMW_GURUMDDS_ID,
+    return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
+  auto service_info = static_cast<rmw_gurumdds_cpp::ServiceInfo *>(rmw_service->data);
+  if (service_info == nullptr) {
+    RMW_SET_ERROR_MSG("invalid service data");
+    return RMW_RET_ERROR;
+  }
 
-  RMW_SET_ERROR_MSG("rmw_service_set_on_new_request_callback not implemented");
-  return RMW_RET_UNSUPPORTED;
+  std::lock_guard<std::mutex> guard(service_info->event_callback_data.mutex);
+  dds_StatusMask mask = dds_DataReader_get_status_changes(service_info->request_reader);
+  dds_ReturnCode_t dds_rc = dds_RETCODE_ERROR;
+
+  if (callback) {
+    size_t unread_count = service_info->count_unread();
+    if (0 < unread_count) {
+      callback(user_data, unread_count);
+    }
+
+    service_info->event_callback_data.callback = callback;
+    service_info->event_callback_data.user_data = user_data;
+    mask |= dds_DATA_AVAILABLE_STATUS;
+    dds_rc = dds_DataReader_set_listener(service_info->request_reader, &service_info->request_listener, mask);
+  } else {
+    service_info->event_callback_data.callback = nullptr;
+    service_info->event_callback_data.user_data = nullptr;
+    mask &= ~dds_DATA_AVAILABLE_STATUS;
+    dds_rc = dds_DataReader_set_listener(service_info->request_reader, &service_info->request_listener, mask);
+  }
+
+  return rmw_gurumdds_cpp::check_dds_ret_code(dds_rc);
 }
 }  // extern "C"
