@@ -14,15 +14,10 @@
 
 #include <map>
 #include <set>
-#include <string>
 
 #include "rcutils/allocator.h"
-#include "rcutils/logging_macros.h"
-#include "rcutils/strdup.h"
-#include "rcutils/types/string_array.h"
 
 #include "rmw/allocators.h"
-#include "rmw/convert_rcutils_ret_to_rmw_ret.h"
 #include "rmw/error_handling.h"
 #include "rmw/get_service_names_and_types.h"
 #include "rmw/impl/cpp/macros.hpp"
@@ -32,7 +27,6 @@
 #include "rmw_gurumdds_cpp/dds_include.hpp"
 #include "rmw_gurumdds_cpp/demangle.hpp"
 #include "rmw_gurumdds_cpp/identifier.hpp"
-#include "rmw_gurumdds_cpp/names_and_types_helpers.hpp"
 #include "rmw_gurumdds_cpp/rmw_context_impl.hpp"
 
 extern "C"
@@ -58,8 +52,8 @@ rmw_get_service_names_and_types(
 
   auto common_ctx = &node->context->impl->common_ctx;
   return common_ctx->graph_cache.get_names_and_types(
-    _demangle_service_from_topic,
-    _demangle_service_type_only,
+    rmw_gurumdds_cpp::demangle_service_from_topic,
+    rmw_gurumdds_cpp::demangle_service_type_only,
     allocator,
     service_names_and_types);
 }
