@@ -1,4 +1,4 @@
-// Copyright 2019 GurumNetworks, Inc.
+// Copyright 2024 GurumNetworks, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,22 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef RMW_GURUMDDS_CPP__RMW_WAIT_HPP_
-#define RMW_GURUMDDS_CPP__RMW_WAIT_HPP_
-
 #include <chrono>
 #include <unordered_map>
 #include <unordered_set>
-#include <utility>
+
+#include "rcpputils/scope_exit.hpp"
 
 #include "rmw/allocators.h"
 #include "rmw/error_handling.h"
-#include "rmw/rmw.h"
 #include "rmw/impl/cpp/macros.hpp"
 
 #include "rmw_gurumdds_cpp/dds_include.hpp"
 #include "rmw_gurumdds_cpp/event_converter.hpp"
-#include "rmw_gurumdds_cpp/types.hpp"
+#include "rmw_gurumdds_cpp/event_info_common.hpp"
+#include "rmw_gurumdds_cpp/event_info_service.hpp"
+#include "rmw_gurumdds_cpp/wait.hpp"
 
 #define CHECK_ATTACH(ret) \
   if (ret == dds_RETCODE_OK) { \
@@ -43,10 +42,12 @@
     return RMW_RET_ERROR; \
   }
 
-rmw_ret_t
-__gather_event_conditions(
+namespace rmw_gurumdds_cpp
+{
+static rmw_ret_t
+gather_event_conditions(
   rmw_events_t * events,
-  std::unordered_set<dds_StatusCondition *> & status_conditions)
+  std::unordered_set<dds_Condition*> & status_conditions)
 {
   RMW_CHECK_ARGUMENT_FOR_NULL(events, RMW_RET_INVALID_ARGUMENT);
   std::unordered_map<dds_StatusCondition *, dds_StatusMask> status_map;
@@ -55,37 +56,45 @@ __gather_event_conditions(
     auto now = static_cast<rmw_event_t *>(events->events[i]);
     RMW_CHECK_ARGUMENT_FOR_NULL(events, RMW_RET_INVALID_ARGUMENT);
 
-    auto event_info = static_cast<GurumddsEventInfo *>(now->data);
-    if (event_info == nullptr) {
+    auto event_info = static_cast<EventInfo *>(now->data);
+    if (nullptr == event_info) {
       RMW_SET_ERROR_MSG("event handle is null");
       return RMW_RET_ERROR;
     }
 
-    dds_StatusCondition * status_condition = event_info->get_statuscondition();
-    if (status_condition == nullptr) {
-      RMW_SET_ERROR_MSG("failed to get status condition");
-      return RMW_RET_ERROR;
+    const auto event_type = now->event_type;
+    if(!is_event_supported(event_type)) {
+      RMW_SET_ERROR_MSG_WITH_FORMAT_STRING("unsupported event: %d", now->event_type);
+      continue;
     }
 
-    if (is_event_supported(now->event_type)) {
-      auto map_pair = status_map.insert(std::make_pair(status_condition, 0));
-      auto it = map_pair.first;
-      status_map[status_condition] = get_status_kind_from_rmw(now->event_type) | it->second;
-    } else {
-      RMW_SET_ERROR_MSG_WITH_FORMAT_STRING("unsupported event: %d", now->event_type);
+    if(event_info->has_callback(event_type)) {
+      dds_GuardCondition * condition = event_info->get_guard_condition(event_type);
+      if (nullptr == condition) {
+        RMW_SET_ERROR_MSG_WITH_FORMAT_STRING("unsupported event: %d", event_type);
+        continue;
+      }
+
+      status_conditions.insert(reinterpret_cast<dds_Condition *>(condition));
     }
+
+    auto& mask = status_map[event_info->get_status_condition()];
+    mask |= get_status_kind_from_rmw(event_type);
   }
 
-  for (auto & map_pair : status_map) {
-    dds_StatusCondition_set_enabled_statuses(map_pair.first, map_pair.second);
-    status_conditions.insert(map_pair.first);
+  for(auto & pair : status_map) {
+    if(pair.second == 0) {
+      continue;
+    }
+
+    dds_StatusCondition_set_enabled_statuses(pair.first, pair.second);
+    status_conditions.insert(reinterpret_cast<dds_Condition *>(pair.first));
   }
 
   return RMW_RET_OK;
 }
 
-rmw_ret_t
-__handle_active_event_conditions(rmw_events_t * events)
+static rmw_ret_t handle_active_event_conditions(rmw_events_t * events)
 {
   if (events == nullptr) {
     return RMW_RET_OK;
@@ -95,20 +104,13 @@ __handle_active_event_conditions(rmw_events_t * events)
     auto now = static_cast<rmw_event_t *>(events->events[i]);
     RMW_CHECK_ARGUMENT_FOR_NULL(events, RMW_RET_INVALID_ARGUMENT);
 
-    auto event_info = static_cast<GurumddsEventInfo *>(now->data);
+    auto event_info = static_cast<EventInfo *>(now->data);
     if (event_info == nullptr) {
       RMW_SET_ERROR_MSG("event handle is null");
       return RMW_RET_ERROR;
     }
 
-    dds_StatusMask mask = event_info->get_status_changes();
-    bool is_active = false;
-
-    if (is_event_supported(now->event_type)) {
-      is_active = ((mask & get_status_kind_from_rmw(now->event_type)) != 0);
-    }
-
-    if (!is_active) {
+    if (!event_info->is_status_changed(now->event_type)) {
       events->events[i] = nullptr;
     }
   }
@@ -116,13 +118,14 @@ __handle_active_event_conditions(rmw_events_t * events)
   return RMW_RET_OK;
 }
 
-rmw_ret_t __detach_condition(
+static rmw_ret_t
+detach_condition(
   dds_WaitSet * dds_wait_set,
   dds_Condition * condition)
 {
   dds_ReturnCode_t dds_return_code = dds_WaitSet_detach_condition(dds_wait_set, condition);
   rmw_ret_t from_dds = check_dds_ret_code(dds_return_code);
-  if (from_dds != RMW_RET_OK) {
+  if (dds_return_code != dds_RETCODE_PRECONDITION_NOT_MET && from_dds != RMW_RET_OK) {
     RMW_SET_ERROR_MSG("failed to detach condition from wait set");
     return from_dds;
   }
@@ -130,9 +133,68 @@ rmw_ret_t __detach_condition(
   return RMW_RET_OK;
 }
 
-template<typename SubscriberInfo, typename ServiceInfo, typename ClientInfo>
+static rmw_ret_t
+wait_w_polling(
+  dds_WaitSet * dds_wait_set,
+  dds_ConditionSeq * active_conditions,
+  const rmw_time_t * wait_timeout
+    ) {
+  uint64_t sec, nsec;
+  bool inf = false;
+  if (wait_timeout != nullptr) {
+    sec = wait_timeout->sec;
+    nsec = wait_timeout->nsec;
+    inf = false;
+  } else {
+    sec = 0;
+    nsec = 0;
+    inf = true;
+  }
+  auto t = std::chrono::steady_clock::now() +
+           std::chrono::nanoseconds(sec * 1000000000ULL + nsec);
+  bool triggered = false;
+
+  while (dds_ConditionSeq_length(active_conditions) > 0) {
+    dds_ConditionSeq_remove(active_conditions, 0);
+  }
+
+  dds_ConditionSeq * conds = dds_ConditionSeq_create(8);
+  dds_WaitSet_get_conditions(dds_wait_set, conds);
+
+  for (uint32_t i = 0; i < dds_ConditionSeq_length(conds); ++i) {
+    dds_Condition * cond = dds_ConditionSeq_get(conds, i);
+    if (cond == nullptr) {
+      continue;
+    }
+
+    if (dds_Condition_get_trigger_value(cond)) {
+      dds_ConditionSeq_add(active_conditions, cond);
+      triggered = true;
+    }
+  }
+
+  for (uint32_t i = 0; (inf || std::chrono::steady_clock::now() <= t) && !triggered; ++i) {
+    if (i >= dds_ConditionSeq_length(conds)) {
+      i = 0;
+    }
+
+    dds_Condition * cond = dds_ConditionSeq_get(conds, i);
+    if (cond == nullptr) {
+      continue;
+    }
+
+    if (dds_Condition_get_trigger_value(cond)) {
+      dds_ConditionSeq_add(active_conditions, cond);
+      triggered = true;
+    }
+  }
+
+  dds_ConditionSeq_delete(conds);
+  return triggered ? RMW_RET_OK : RMW_RET_TIMEOUT;
+}
+
 rmw_ret_t
-__rmw_wait(
+wait(
   const char * implementation_identifier,
   rmw_subscriptions_t * subscriptions,
   rmw_guard_conditions_t * guard_conditions,
@@ -142,70 +204,63 @@ __rmw_wait(
   rmw_wait_set_t * wait_set,
   const rmw_time_t * wait_timeout)
 {
-  (void)events;
-  struct atexit_t
-  {
-    ~atexit_t()
-    {
-      if (wait_set == nullptr) {
-        RMW_SET_ERROR_MSG("wait set handle is null");
-        return;
-      }
+  RCUTILS_CAN_RETURN_WITH_ERROR_OF(RMW_RET_INVALID_ARGUMENT);
+  RCUTILS_CAN_RETURN_WITH_ERROR_OF(RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
+  auto atexit = rcpputils::make_scope_exit([&]() {
+    if (wait_set == nullptr) {
+      RMW_SET_ERROR_MSG("wait set handle is null");
+      return;
+    }
 
-      RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
-        wait set handle, wait_set->implementation_identifier,
-        implementation_identifier, return )
+    RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
+      wait set handle, wait_set->implementation_identifier,
+      implementation_identifier, return );
 
-      GurumddsWaitSetInfo * wait_set_info = static_cast<GurumddsWaitSetInfo *>(wait_set->data);
-      if (wait_set_info == nullptr) {
-        RMW_SET_ERROR_MSG("WaitSet implementation struct is null");
-        return;
-      }
+    WaitSetInfo * wait_set_info = static_cast<WaitSetInfo *>(wait_set->data);
+    if (wait_set_info == nullptr) {
+      RMW_SET_ERROR_MSG("WaitSet implementation struct is null");
+      return;
+    }
 
-      dds_WaitSet * dds_wait_set = static_cast<dds_WaitSet *>(wait_set_info->wait_set);
-      if (dds_wait_set == nullptr) {
-        RMW_SET_ERROR_MSG("DDS wait set handle is null");
-        return;
-      }
+    dds_WaitSet * dds_wait_set = static_cast<dds_WaitSet *>(wait_set_info->wait_set);
+    if (dds_wait_set == nullptr) {
+      RMW_SET_ERROR_MSG("DDS wait set handle is null");
+      return;
+    }
 
-      dds_ConditionSeq * attached_conditions =
-        static_cast<dds_ConditionSeq *>(wait_set_info->attached_conditions);
-      if (attached_conditions == nullptr) {
-        RMW_SET_ERROR_MSG("DDS condition sequence handle is null");
-        return;
-      }
+    dds_ConditionSeq * attached_conditions =
+      static_cast<dds_ConditionSeq *>(wait_set_info->attached_conditions);
+    if (attached_conditions == nullptr) {
+      RMW_SET_ERROR_MSG("DDS condition sequence handle is null");
+      return;
+    }
 
-      dds_ReturnCode_t ret = dds_WaitSet_get_conditions(dds_wait_set, attached_conditions);
+    dds_ReturnCode_t ret = dds_WaitSet_get_conditions(dds_wait_set, attached_conditions);
+    if (ret != dds_RETCODE_OK) {
+      RMW_SET_ERROR_MSG("failed to get attached conditions for wait set");
+      return;
+    }
+
+    const uint32_t condition_seq_length = dds_ConditionSeq_length(attached_conditions);
+    for (uint32_t i = 0; i < condition_seq_length; ++i) {
+      ret = dds_WaitSet_detach_condition(
+        dds_wait_set, dds_ConditionSeq_get(attached_conditions, i));
       if (ret != dds_RETCODE_OK) {
-        RMW_SET_ERROR_MSG("failed to get attached conditions for wait set");
-        return;
-      }
-
-      for (uint32_t i = 0; i < dds_ConditionSeq_length(attached_conditions); ++i) {
-        ret = dds_WaitSet_detach_condition(
-          dds_wait_set, dds_ConditionSeq_get(attached_conditions, i));
-        if (ret != dds_RETCODE_OK) {
-          RMW_SET_ERROR_MSG("failed to detach condition from wait set");
-        }
-      }
-
-      while (dds_ConditionSeq_length(attached_conditions) > 0) {
-        dds_ConditionSeq_remove(attached_conditions, 0);
+        RMW_SET_ERROR_MSG("failed to detach condition from wait set");
       }
     }
-    rmw_wait_set_t * wait_set = nullptr;
-    const char * implementation_identifier = nullptr;
-  } atexit;
 
-  atexit.wait_set = wait_set;
-  atexit.implementation_identifier = implementation_identifier;
+    while (dds_ConditionSeq_length(attached_conditions) > 0) {
+      dds_ConditionSeq_remove(attached_conditions, 0);
+    }
+  });
 
   RMW_CHECK_ARGUMENT_FOR_NULL(wait_set, RMW_RET_INVALID_ARGUMENT);
   RMW_CHECK_TYPE_IDENTIFIERS_MATCH(
     wait set handle, wait_set->implementation_identifier,
     implementation_identifier, return RMW_RET_INCORRECT_RMW_IMPLEMENTATION);
 
-  GurumddsWaitSetInfo * wait_set_info = static_cast<GurumddsWaitSetInfo *>(wait_set->data);
+  WaitSetInfo * wait_set_info = static_cast<WaitSetInfo *>(wait_set->data);
   if (wait_set_info == nullptr) {
     RMW_SET_ERROR_MSG("WaitSet implementation struct is null");
     return RMW_RET_ERROR;
@@ -226,7 +281,7 @@ __rmw_wait(
 
   if (subscriptions != nullptr) {
     for (size_t i = 0; i < subscriptions->subscriber_count; ++i) {
-      SubscriberInfo * subscriber_info =
+      auto * subscriber_info =
         static_cast<SubscriberInfo *>(subscriptions->subscribers[i]);
       if (subscriber_info == nullptr) {
         RMW_SET_ERROR_MSG("subscriber info handle is null");
@@ -245,23 +300,21 @@ __rmw_wait(
     }
   }
 
-  std::unordered_set<dds_StatusCondition *> status_conditions;
+  std::unordered_set<dds_Condition *> status_conditions;
 
-  rmw_ret_t ret_code = __gather_event_conditions(events, status_conditions);
+  rmw_ret_t ret_code = gather_event_conditions(events, status_conditions);
   if (ret_code != RMW_RET_OK) {
     return ret_code;
   }
 
   for (auto status_condition : status_conditions) {
-    dds_ReturnCode_t ret = dds_WaitSet_attach_condition(
-      dds_wait_set,
-      reinterpret_cast<dds_Condition *>(status_condition));
+    dds_ReturnCode_t ret = dds_WaitSet_attach_condition(dds_wait_set, status_condition);
     CHECK_ATTACH(ret);
   }
 
   if (guard_conditions != nullptr) {
     for (size_t i = 0; i < guard_conditions->guard_condition_count; ++i) {
-      dds_GuardCondition * guard_condition =
+      auto * guard_condition =
         static_cast<dds_GuardCondition *>(guard_conditions->guard_conditions[i]);
       if (guard_condition == nullptr) {
         RMW_SET_ERROR_MSG("guard condition handle is null");
@@ -276,7 +329,7 @@ __rmw_wait(
 
   if (services != nullptr) {
     for (size_t i = 0; i < services->service_count; ++i) {
-      ServiceInfo * service_info = static_cast<ServiceInfo *>(services->services[i]);
+      auto * service_info = static_cast<ServiceInfo *>(services->services[i]);
       if (service_info == nullptr) {
         RMW_SET_ERROR_MSG("service info handle is null");
         return RMW_RET_ERROR;
@@ -296,7 +349,7 @@ __rmw_wait(
 
   if (clients != nullptr) {
     for (size_t i = 0; i < clients->client_count; ++i) {
-      ClientInfo * client_info = static_cast<ClientInfo *>(clients->clients[i]);
+      auto * client_info = static_cast<ClientInfo *>(clients->clients[i]);
       if (client_info == nullptr) {
         RMW_SET_ERROR_MSG("client info handle is null");
         return RMW_RET_ERROR;
@@ -315,17 +368,19 @@ __rmw_wait(
   }
 
   rmw_ret_t rret = RMW_RET_OK;
-
-  const char * env_name = "RMW_GURUMDDS_WAIT_USE_POLLING";
-  char * env_value = nullptr;
-  bool use_polling = false;
-
-  env_value = getenv(env_name);
-  if (env_value != nullptr) {
-    use_polling = (strcmp(env_value, "1") == 0);
+  static constexpr const char * env_name = "RMW_GURUMDDS_WAIT_USE_POLLING";
+  static bool initialized_polling = false;
+  static bool polling = false;
+  if(!initialized_polling) {
+    initialized_polling = true;
+    char * env_value = nullptr;
+    env_value = getenv(env_name);
+    if (nullptr != env_value) {
+      polling = (strcmp(env_value, "1") == 0);
+    }
   }
 
-  if (!use_polling) {  // Default: use dds_WaitSet_wait()
+  if (!polling) {  // Default: use dds_WaitSet_wait()
     dds_Duration_t timeout;
     if (wait_timeout == nullptr) {
       timeout.sec = dds_DURATION_INFINITE_SEC;
@@ -345,62 +400,10 @@ __rmw_wait(
       rret = RMW_RET_TIMEOUT;
     }
   } else {  // use polilng
-    uint64_t sec, nsec;
-    bool inf = false;
-    if (wait_timeout != nullptr) {
-      sec = wait_timeout->sec;
-      nsec = wait_timeout->nsec;
-      inf = false;
-    } else {
-      sec = 0;
-      nsec = 0;
-      inf = true;
-    }
-    auto t = std::chrono::steady_clock::now() +
-      std::chrono::nanoseconds(sec * 1000000000ULL + nsec);
-    bool triggered = false;
-
-    while (dds_ConditionSeq_length(active_conditions) > 0) {
-      dds_ConditionSeq_remove(active_conditions, 0);
-    }
-
-    dds_ConditionSeq * conds = dds_ConditionSeq_create(8);
-    dds_WaitSet_get_conditions(dds_wait_set, conds);
-
-    for (uint32_t i = 0; i < dds_ConditionSeq_length(conds); ++i) {
-      dds_Condition * cond = dds_ConditionSeq_get(conds, i);
-      if (cond == NULL) {
-        continue;
-      }
-
-      if (dds_Condition_get_trigger_value(cond) == true) {
-        dds_ConditionSeq_add(active_conditions, cond);
-        triggered = true;
-      }
-    }
-
-    for (uint32_t i = 0; (inf || std::chrono::steady_clock::now() <= t) && !triggered; ++i) {
-      if (i >= dds_ConditionSeq_length(conds)) {
-        i = 0;
-      }
-
-      dds_Condition * cond = dds_ConditionSeq_get(conds, i);
-      if (cond == NULL) {
-        continue;
-      }
-
-      if (dds_Condition_get_trigger_value(cond) == true) {
-        dds_ConditionSeq_add(active_conditions, cond);
-        triggered = true;
-      }
-    }
-    dds_ConditionSeq_delete(conds);
-
-    if (!triggered) {
-      rret = RMW_RET_TIMEOUT;
-    }
+    rret = wait_w_polling(dds_wait_set, active_conditions, wait_timeout);
   }
 
+  const uint32_t active_cond_length = dds_ConditionSeq_length(active_conditions);
   if (subscriptions != nullptr) {
     for (size_t i = 0; i < subscriptions->subscriber_count; ++i) {
       SubscriberInfo * subscriber_info =
@@ -417,20 +420,19 @@ __rmw_wait(
       }
 
       uint32_t j = 0;
-      for (; j < dds_ConditionSeq_length(active_conditions); ++j) {
+      for (; j < active_cond_length; ++j) {
         if (
           dds_ConditionSeq_get(active_conditions, j) ==
-          reinterpret_cast<dds_Condition *>(read_condition))
-        {
+          reinterpret_cast<dds_Condition *>(read_condition)) {
           break;
         }
       }
 
-      if (j >= dds_ConditionSeq_length(active_conditions)) {
-        subscriptions->subscribers[i] = 0;
+      if (j >= active_cond_length) {
+        subscriptions->subscribers[i] = nullptr;
       }
 
-      rmw_ret_t rmw_ret_code = __detach_condition(
+      rmw_ret_t rmw_ret_code = detach_condition(
         dds_wait_set, reinterpret_cast<dds_Condition *>(read_condition));
       if (rmw_ret_code != RMW_RET_OK) {
         return rmw_ret_code;
@@ -440,7 +442,7 @@ __rmw_wait(
 
   if (guard_conditions != nullptr) {
     for (size_t i = 0; i < guard_conditions->guard_condition_count; ++i) {
-      dds_Condition * condition =
+      auto * condition =
         static_cast<dds_Condition *>(guard_conditions->guard_conditions[i]);
       if (condition == nullptr) {
         RMW_SET_ERROR_MSG("condition handle is null");
@@ -448,9 +450,9 @@ __rmw_wait(
       }
 
       uint32_t j = 0;
-      for (; j < dds_ConditionSeq_length(active_conditions); ++j) {
+      for (; j < active_cond_length; ++j) {
         if (dds_ConditionSeq_get(active_conditions, j) == condition) {
-          dds_GuardCondition * guard = reinterpret_cast<dds_GuardCondition *>(condition);
+          auto * guard = reinterpret_cast<dds_GuardCondition *>(condition);
           dds_ReturnCode_t ret = dds_GuardCondition_set_trigger_value(guard, false);
           if (ret != dds_RETCODE_OK) {
             RMW_SET_ERROR_MSG("failed to set trigger value");
@@ -460,11 +462,11 @@ __rmw_wait(
         }
       }
 
-      if (j >= dds_ConditionSeq_length(active_conditions)) {
-        guard_conditions->guard_conditions[i] = 0;
+      if (j >= active_cond_length) {
+        guard_conditions->guard_conditions[i] = nullptr;
       }
 
-      rmw_ret_t rmw_ret_code = __detach_condition(dds_wait_set, condition);
+      rmw_ret_t rmw_ret_code = detach_condition(dds_wait_set, condition);
       if (rmw_ret_code != RMW_RET_OK) {
         return rmw_ret_code;
       }
@@ -473,7 +475,7 @@ __rmw_wait(
 
   if (services != nullptr) {
     for (size_t i = 0; i < services->service_count; ++i) {
-      ServiceInfo * service_info = static_cast<ServiceInfo *>(services->services[i]);
+      auto * service_info = static_cast<ServiceInfo *>(services->services[i]);
       if (service_info == nullptr) {
         RMW_SET_ERROR_MSG("service info handle is null");
         return RMW_RET_ERROR;
@@ -486,7 +488,7 @@ __rmw_wait(
       }
 
       uint32_t j = 0;
-      for (; j < dds_ConditionSeq_length(active_conditions); ++j) {
+      for (; j < active_cond_length; ++j) {
         if (
           dds_ConditionSeq_get(active_conditions, j) ==
           reinterpret_cast<dds_Condition *>(read_condition))
@@ -495,11 +497,11 @@ __rmw_wait(
         }
       }
 
-      if (j >= dds_ConditionSeq_length(active_conditions)) {
-        services->services[i] = 0;
+      if (j >= active_cond_length) {
+        services->services[i] = nullptr;
       }
 
-      rmw_ret_t rmw_ret_code = __detach_condition(
+      rmw_ret_t rmw_ret_code = detach_condition(
         dds_wait_set, reinterpret_cast<dds_Condition *>(read_condition));
       if (rmw_ret_code != RMW_RET_OK) {
         return rmw_ret_code;
@@ -509,7 +511,7 @@ __rmw_wait(
 
   if (clients != nullptr) {
     for (size_t i = 0; i < clients->client_count; ++i) {
-      ClientInfo * client_info = static_cast<ClientInfo *>(clients->clients[i]);
+      auto * client_info = static_cast<ClientInfo *>(clients->clients[i]);
       if (client_info == nullptr) {
         RMW_SET_ERROR_MSG("client info handle is null");
         return RMW_RET_ERROR;
@@ -522,7 +524,7 @@ __rmw_wait(
       }
 
       uint32_t j = 0;
-      for (; j < dds_ConditionSeq_length(active_conditions); ++j) {
+      for (; j < active_cond_length; ++j) {
         if (
           dds_ConditionSeq_get(active_conditions, j) ==
           reinterpret_cast<dds_Condition *>(read_condition))
@@ -531,11 +533,11 @@ __rmw_wait(
         }
       }
 
-      if (j >= dds_ConditionSeq_length(active_conditions)) {
-        clients->clients[i] = 0;
+      if (j >= active_cond_length) {
+        clients->clients[i] = nullptr;
       }
 
-      rmw_ret_t rmw_ret_code = __detach_condition(
+      rmw_ret_t rmw_ret_code = detach_condition(
         dds_wait_set, reinterpret_cast<dds_Condition *>(read_condition));
       if (rmw_ret_code != RMW_RET_OK) {
         return rmw_ret_code;
@@ -543,14 +545,11 @@ __rmw_wait(
     }
   }
 
-  {
-    rmw_ret_t rmw_ret_code = __handle_active_event_conditions(events);
-    if (rmw_ret_code != RMW_RET_OK) {
-      return rmw_ret_code;
-    }
+  rmw_ret_t rmw_ret_code = handle_active_event_conditions(events);
+  if (rmw_ret_code != RMW_RET_OK) {
+    return rmw_ret_code;
   }
 
   return rret;
 }
-
-#endif  // RMW_GURUMDDS_CPP__RMW_WAIT_HPP_
+} // namespace rmw_gurumdds_cpp
